--- ERATAG_REPORT.txt
@@ -1,69 +1,0 @@
-ERA-TAG DECORRELATION PASS — analysis_configurations/tau/config.py
-====================================================================
-Branch: audit-run3-htautau-eratags (nested TauAnalysis-CROWN repo)
-Scope: tasks "use era tag for uncorrelated uncertainties as indicated in
-the twikis of the objects", "decorrelate z pt uncertainties per era", and
-"tau variation uncorrelated by era per the TauPOG (non-CMS-TWiki) reference".
-
-Sources used: systematics_master.txt, Conventions_-_CMS_Common_Analysis_
-Tools_Documentation.txt, Introduction_-_TauPOG_Wiki.txt, DY_kinematic_
-reweighting_-_CMS_Higgs_to_leptons_and_rare_decays.txt (all in
-scratchpad/twikis_txt/), plus the prior AUDIT_REPORT_tau.txt and
-AUDIT_REPORT_zpt.txt.
-
-WHAT WAS DONE
---------------
-Added a module-level ERA_MAP (matching jet_variations.py's ERA_MAP /
-systematics_master.yml era tokens: 2022preEE->2022, 2022postEE->2022EE,
-2023preBPix->2023, 2023postBPix->2023BPix, else identity) and
-`era_tag = ERA_MAP[era]` at the top of build_config. Embedded era_tag
-into the shift `name=` string (not the correctionlib key/shift_map, which
-stays unchanged) for every uncorrelated-by-era uncertainty I could
-confirm against either an explicit era regex group in systematics_master
-or an explicit "uncorrelated across years/eras" sentence in the TauPOG
-Wiki:
-
-  muonEsScale, muonEsReso, eleEsReso, eleEsScale, electronIdSF, PileUp,
-  zPtReweightWeight_unc{n} (both NNLO/powheg and NLO/amcatnlo loops),
-  tauEsDM{dm}, tauEsDM{dm}pT{pt}, tauEleFakeEsDM{dm}Barrel/Endcap,
-  tauMuFakeEsDM{dm}, vsEleDM{dm}Barrel/Endcap (era>=2022), vsMuWheel{wheel}
-  (era>=2022 branch only).
-
-Also fixed vsJetStat1/2{era}DM{dm}, vsJetSystTES{era}DM{dm}, and
-vsJetSyst{era} (added by a prior tau-SF audit) to use the mapped era_tag
-token instead of the raw era string, so "2022postEE" becomes "2022EE" in
-the shift name, matching every other era-tagged shift in this codebase
-and the systematics_master token convention. This is a spelling fix only
-— era was already embedded, so decorrelation behavior is unchanged.
-
-Full details, including exact rationale/twiki citation per shift and a
-list of what was deliberately NOT touched (Run2 vsEle/vsMu names, MET
-recoil naming, prefiring source ambiguity, CMS_htt_zptReweight's missing
-CADI id, and the fact that the CMS_* prefix rename itself was not applied
-to any of these — only era_tag suffixes were added to the existing legacy
-names), are appended to shift_rename_log.txt under "Follow-up pass:
-era-tag decorrelation".
-
-VERIFICATION
--------------
-`python3 -c "import ast; ast.parse(open('config.py').read())"` passes.
-Attempted a `build_config()` smoke run via direct import for all Run2/
-Run3 eras; hit the same class of pre-existing SampleRuleConfigurationError/
-SampleConfigurationError as documented in AUDIT_REPORT_tau.txt (sample-
-type list bookkeeping in the test harness invocation, unrelated to this
-change — reproduces identically for eras/branches this pass did not
-touch). Did not chase further; all changes here are structural renames of
-`name=` f-strings only, with no change to shift_map/shift_key logic.
-
-NOT DONE / LEFT OPEN
-----------------------
-- The CMS_* prefix rename for the shifts touched here (tracked separately
-  in shift_rename_log.txt's earlier sections) was not applied in this
-  pass, to keep the diff focused on decorrelation.
-- Run2 (<2022) vsEle/vsMu/vsJetTau shift names: out of scope (Run3-focused
-  pass), and systematics_master doesn't list an era-dependent group for
-  VSe/VSmu fake-tau ID SF at all, so this is not unambiguously required
-  by the generic convention.
-- metRecoilResponse/metRecoilResolution era-tagging: intentionally left
-  for the separate recoil-producer-consolidation task to avoid stepping
-  on that work.
