--- RECOIL_REPORT.txt
@@ -1,95 +1,0 @@
-Recoil producer merge/refactor audit
-=====================================
-
-Scope: producers/met.py (Run3 recoil correction + uncertainty producers) and
-config.py's recoil-related config parameters. Read-only reference material:
-AUDIT_REPORT_met.txt (prior fix history), the HLepRare recoil twiki extract,
-and (read-only, out of scope for edits) the outer CROWN repo's src/met.cxx
-and include/met.hxx for the underlying met::RecoilCorrection signature.
-
-Current structure (before this change)
----------------------------------------
-Two separate Producer objects both call the same C++ function
-`met::RecoilCorrection(...)`, with the identical positional argument order:
-
-- ApplyRecoilCorrections: method = "{recoil_method}" (config-templated,
-  normally "QuantileMapHist"), variation = "{recoil_variation}", toggle =
-  "{applyRecoilCorrections}". Input: puppimet_p4_leptoncorrected. Output:
-  puppimet_p4_recoilcorrected.
-- ApplyRecoilUncertainty: method hardcoded to the literal "Uncertainty",
-  variation = "{recoil_uncertainty_variation}", toggle =
-  "{applyRecoilUncertainty}". Input: puppimet_p4_recoilcorrected (i.e. the
-  OUTPUT of ApplyRecoilCorrections - this is the audit's correctness fix:
-  the uncertainty must be evaluated on the already nominally-recoil-corrected
-  MET, per the HLepRare doc). Output: puppimet_p4_recoiluncertaintycorrected.
-
-Both were already chained in the right order as consecutive subproducers
-inside MetCorrections and MetCorrections_v12 (each ProducerGroup listed them
-back-to-back), so the correctness fix was already structurally sound - just
-duplicated across two ProducerGroups (v15 and v12 pipelines).
-
-Why a Producer-level (single call / shared C++ signature) merge is NOT
-advisable
-------------------------------------------------------------------------
-Although the call signature is structurally identical, CROWN's Producer
-`call` string is resolved by substituting config-parameter names verbatim at
-code-gen time (no Python branching inside a single call template). The two
-steps intentionally use DIFFERENT config parameter names
-(recoil_method/recoil_variation/applyRecoilCorrections vs
-recoil_uncertainty_variation/applyRecoilUncertainty, plus a literal
-"Uncertainty" method) precisely so that:
-  1. the nominal correction always runs, unconditionally, in "QuantileMapHist"
-     mode, and
-  2. the uncertainty step can be turned on/off and steered independently via
-     config.py's shift_key mechanism (metRecoilResponse/metRecoilResolution
-     shifts target met.ApplyRecoilUncertainty specifically, see config.py
-     ~line 2454).
-Collapsing these into one Producer/call would either reuse one config key for
-two purposes (breaking independent shift control) or require conditional
-Python logic to pick the call string per-instantiation, which is not how this
-declarative Producer/ProducerGroup layer works elsewhere in the repo. It would
-not reduce real duplication (the C++ call text itself is only ~1 line per
-producer) and would risk breaking the shift-targeting mechanism. Also,
-ApplyRecoilCorrections is used standalone (without the uncertainty step) in
-doublemuon_controlregion.py's `mm` control region path, so it must remain an
-independently addressable Producer.
-
-What WAS merged
-----------------
-The two-producer chain (ApplyRecoilCorrections -> ApplyRecoilUncertainty) was
-previously repeated verbatim in both MetCorrections and MetCorrections_v12.
-Bundled this pair into a new nested ProducerGroup, RecoilCorrectionRun3
-(producers/met.py), following the same nested-ProducerGroup pattern already
-used elsewhere in this repo (e.g. muons.py's GoodMuonsWithDzDxyCuts, jets.py's
-JetEnergyCorrection). MetCorrections and MetCorrections_v12 now reference
-RecoilCorrectionRun3 instead of listing both producers separately, removing
-the duplication while preserving:
-  - execution order (nominal correction still always runs before the
-    uncertainty step),
-  - all existing output quantity names (puppimet_p4_recoilcorrected,
-    puppimet_p4_recoiluncertaintycorrected unchanged - grepped the whole repo
-    for all consumers, including producers/pairquantities.py and
-    doublemuon_controlregion.py, none renamed),
-  - config.py's shift-targeting of met.ApplyRecoilUncertainty directly (the
-    shift_key override at config.py ~line 2454 still addresses the individual
-    producer object, which still exists and is unchanged; wrapping it in a
-    ProducerGroup for the "always chain these two" use case does not affect
-    referencing it directly elsewhere),
-  - doublemuon_controlregion.py's standalone use of met.ApplyRecoilCorrections
-    without the uncertainty step (untouched).
-
-Verification
-------------
-`python3 -c "import ast; ast.parse(open('producers/met.py').read())"` passes.
-Grepped the whole repo for all recoil output quantity usages
-(puppimet_p4_recoilcorrected, puppimet_p4_recoiluncertaintycorrected,
-pfmet_p4_recoilcorrected, met.ApplyRecoilCorrections, met.ApplyRecoilUncertainty)
-before and after - no renames, no consumer breakage.
-
-Note (out of scope, not touched): producers/pairquantities.py's MT/PZeta/etc.
-producers take q.puppimet_p4_recoilcorrected directly as MET input rather than
-the uncertainty-corrected or unclustered-corrected MET - meaning recoil
-response/resolution shifts do not propagate into those pair-quantity
-calculations. This looks like it might be worth a separate follow-up audit,
-but is unrelated to the merge/duplication question asked here and was left
-unchanged.
