--- TRIGGER2_REPORT.txt
@@ -1,81 +1,0 @@
-TRIGGER FOLLOW-UP AUDIT (tasks #5 + #7) — mutau/eltau/singletau filterbits, ditau verification
-================================================================================================
-Branch: audit-run3-htautau-trigger2 (nested TauAnalysis-CROWN repo)
-Continues AUDIT_REPORT_trigger.txt's "GAPS / COULD NOT FULLY VERIFY" section, using the full
-extracted TauTrigger_CMS_TWiki.txt (4002 lines) as reference.
-
-DOCUMENTATION SOURCE / PRECEDENCE CHECK
-------------------------------------------
-Per the task instructions, checked the reference doc's own header/URL before trusting it as
-higher-precedence. Line 10 of TauTrigger_CMS_TWiki.txt reads:
-  "TWiki> CMS Web>Tau>TauTrigger (2026-04-24, BradenAllmond)"
-This confirms the doc genuinely IS the CMS-internal TWiki (twiki.cern.ch, "CMS Web" space) —
-matching its filename, NOT a TauPOG-hosted page (e.g. tau-wiki.docs.cern.ch, which is what the
-sibling tau-ID audit used and which would have outranked it per the stated precedence rule). No
-TauPOG-hosted trigger-specific page was available for this task, so this CMS TWiki was used as
-the sole source at its normal (lower) precedence tier — there was no conflicting higher-tier doc
-to reconcile against for the trigger topic specifically.
-
-ISSUES FOUND AND FIXED (tau_triggersetup.py)
-----------------------------------------------
-Root cause common to all three: the twiki's per-year trigger tables report DIFFERENT NanoAOD
-filterbit indices for the *same* HLT path name across NanoAOD production campaigns (V12 for
-2022preEE/2022postEE/2023preBPix/2023postBPix vs V14/V15 for 2024/2025). A single filterbit
-value had been reused across all of 2022preEE..2024, matching neither campaign correctly in
-two of the three cases:
-
-1. mutau_cross_trigger (HLT_IsoMu20_eta2p1_LooseDeepTauPFTauHPS27_eta2p1_CrossL1), tau-leg
-   p2_filterbit: was a single bit "9" for 2022preEE..2024. Doc gives "3, 9" (V12, 2022-2023)
-   and "3, 13" (V14/V15, 2024) — the old value matched neither (missing the common bit 3
-   for 2022-2023, and wrong second bit entirely for 2024). Fixed per-era.
-2. eltau_cross_trigger (HLT_Ele24_eta2p1_WPTight_Gsf_LooseDeepTauPFTauHPS30_eta2p1_CrossL1),
-   tau-leg p2_filterbit: was a single bit "3" for 2022preEE..2024 (only the common bit, missing
-   the ETau-specific one). Doc gives "3, 8" (V12) and "3, 12" (V14/V15, 2024). Fixed per-era.
-3. singletau_trigger_leading/trailing (HLT_LooseDeepTauPFTauHPS180_L2NN_eta2p1): the ORIGINAL
-   trigger audit (AUDIT_REPORT_trigger.txt) had already fixed this from a Run2 placeholder to
-   this Run3 path, using filterbit "3, 9" uniformly for 2022preEE..2024 — but "3, 9" is only
-   correct for the V14/V15 2024 table; the V12 tables (2022, 2023) give "3, 10". Corrected
-   2022preEE/2022postEE/2023preBPix/2023postBPix to "3, 10", left 2024 at "3, 9".
-
-PNet-era (2025/2026) filterbits for mutau/eltau/singletau/doubletau, and the 2022-2023
-doubletau_trigger DeepTau HPS filterbits ("3, 7"), were already correct against the doc —
-no change needed there.
-
-Verified src/triggers.cxx's matchParticle(): trigger_bit_values entries are bit *indices*
-tested individually with AND semantics (IntBits.test(bit_value) for each), confirming
-multi-bit filterbit strings like "3, 9" are handled correctly and my fixes are semantically
-right (all listed bits must be set, matching the doc's intent).
-
-Commit: 3915138 "Fix mutau/eltau/singletau tau-leg filterbits per NanoAOD version"
-
-DITAU TRIGGER VERIFICATION (task #7) — VERIFIED CORRECT, NO CHANGES NEEDED
-----------------------------------------------------------------------------
-- doubletau_trigger HLT paths/filterbits per era already match the doc exactly: 2022-2023
-  DeepTau HPS35 "3, 7"; 2024-2026 PNet Medium "1, 4, 11". No bug found (unlike mutau/eltau).
-- DoubleTauTriggerLeg1SF/Leg2SF (producers/scalefactors.py) both call
-  physicsobject::tau::scalefactor::Trigger() with trigger_name="ditau", wp="Medium",
-  corrtype="sf", against tau_sf_file's "tau_trigger" correction — config values live in
-  tau_triggersetup.py's doubletau_trigger_leg1_sf/leg2_sf vec_configs (config.py has no
-  doubletau_trigger_* keys at all; this is intentional, not a gap).
-- Inspected the deployed tau.json.gz on cvmfs directly for 2022preEE (V12-era) and 2024
-  (V15-era) tau_sf_file paths: correction "tau_trigger" inputs = [pt, dm, trigtype, wp,
-  corrtype, syst]; "ditau" is a valid trigtype key in both; "Medium" is a valid wp key in
-  both (2024's json only exposes Medium at all — matches the code's "Medium"-only choice
-  everywhere); "sf" is a valid corrtype key. This matches exactly the argument order/values
-  in src/taus.cxx's Trigger(pt, decay_mode, trigger_flag,...) overload
-  (evaluate({pt, dm_or_-1, trigger_name, wp, corr_type, variation})) — no mismatch.
-- Both legs correctly use the SAME "ditau" trigtype (each tau leg gets its own SF lookup off
-  its own pt/dm), consistent with the doc's statement that ditau trigger SF is a tau-leg-only
-  correction with no separate leg1/leg2 split in the json itself.
-- Era-decorrelated Up/Down shift wiring for the ditau trigger SF (the "_variation" loop
-  further down tau_triggersetup.py, ~line 1936) correctly re-supplies era-specific
-  doubletau_trigger_flag/sf_name per the same EraModifier pattern as the nominal case.
-
-No changes made in this section; this closes out the "6. ditau_trigger_wp ... applied
-uniformly" verification item from AUDIT_REPORT_trigger.txt with a positive confirmation
-down to the json/C++ argument level, not just the WP-name level checked previously.
-
-FILES CHANGED
---------------
-- tau_triggersetup.py (commit 3915138, plus the pre-existing muon_sf_variation fix commit
-  1e56955 already present on this branch before I started)
