+++ config.py
@@ -1,4 +1,4 @@
-from __future__ import annotations  # needed for type annotations in > python 3.7
+from __future__ import annotations
 
 from typing import List
 from itertools import product
@@ -14,34 +14,14 @@ from .producers import pairselection as pairselection
 from .producers import scalefactors as scalefactors
 from .producers import taus as taus
 from .producers import triggers as triggers
-from .quantities import nanoAODv15, nanoAODv9
+from .quantities import nanoAODv15
 from .quantities import output as q
 from .tau_triggersetup import add_diTauTriggerSetup
-from .jet_variations import add_jetVariations
+from .variations import add_Variations
 from .tau_embedding_settings import setup_embedding
 from code_generation.configuration import Configuration
 from code_generation.modifiers import EraModifier, SampleModifier
 from code_generation.rules import AppendProducer, RemoveProducer, ReplaceProducer
-from code_generation.systematics import SystematicShift, SystematicShiftByQuantity
-from .scripts.CROWNWrapper import defaults, get_adjusted_add_shift_SystematicShift
-
-# Era tokens matching the CMS Common Analysis Tools systematics_master.yml
-# convention (e.g. CMS_res_e_2022EE, CMS_scale_t_..._2023BPix), used to tag
-# shift names for uncertainties that must be decorrelated across eras.
-ERA_MAP = {
-    "2016preVFP": "2016preVFP",
-    "2016postVFP": "2016postVFP",
-    "2017": "2017",
-    "2018": "2018",
-    "2022preEE": "2022",
-    "2022postEE": "2022EE",
-    "2023preBPix": "2023",
-    "2023postBPix": "2023BPix",
-    "2024": "2024",
-    "2025": "2025",
-    "2026": "2026",
-}
-
 
 def build_config(
     era: str,
@@ -63,8 +43,6 @@ def build_config(
     )
 
     measure_btag_efficiency = False
-
-    era_tag = ERA_MAP[era]
 
     ###########################
     ####### Parameters ########
@@ -255,44 +233,31 @@ def build_config(
             "jet_eta_1": 2.5,
             "jet_eta_2": EraModifier(
                 {
-                    "2016preVFP": 4.7,
-                    "2016postVFP": 4.7,
-                    "2017": 4.7,
-                    "2018": 4.7,
-                    "2022preEE": 4.7, # should be 5 but max cut is 4.7 anyway https://indico.cern.ch/event/1624984/contributions/6896120/attachments/3208048/5713070/20260127_JetMET_PerformanceRun3_HIGMeeting.pdf
-                    "2022postEE": 4.7,
-                    "2023preBPix": 4.7,
-                    "2023postBPix": 4.7,
-                    "2024": 3,
-                    "2025": 3, # same treatment as 2024 for jet eta spikes
-                    "2026": 3, # same treatment as 2024 for jet eta spikes
-                }
+                    "2024": 3,  # jet eta spikes
+                    "2025": 3,  # same treatment as 2024 for jet eta spikes
+                    "2026": 3,  # same treatment as 2024 for jet eta spikes
+                },
+                default=4.7,
             ),
             "jet_eta_3": 4.7,
             "jet_id": 2,  #2==pass tight ID and fail tightLepVeto, 6== pass tight and pass tightLepVeto, new minimal selection https://cms-talk.web.cern.ch/t/updated-jet-selection-criterion-for-jet-veto-map/130527
             # bjet selection -> need to be in global
             "min_bjet_pt": 20,
             "max_bjet_eta": 2.5,
-            "btag_cut": EraModifier( ## values from the wiki for a medium wp https://btv-wiki.docs.cern.ch/ScaleFactors
+            "btag_cut": EraModifier(  ## values from the wiki for a medium wp https://btv-wiki.docs.cern.ch/ScaleFactors
                 {
-                    # wp for deepJEt
+                    # wp for deepJet
                     "2016preVFP": 0.2598,  # taken from https://twiki.cern.ch/twiki/bin/view/CMS/BtagRecommendation106XUL16preVFP
                     "2016postVFP": 0.2489,  # taken from https://twiki.cern.ch/twiki/bin/view/CMS/BtagRecommendation106XUL16postVFP
                     "2017": 0.3040,
                     "2018": 0.2783,
-                    # "2022preEE": 0.3086, 
-                    # "2022postEE": 0.3196,
-                    # "2023preBPix": 0.2431,
-                    # "2023postBPix": 0.2435,
-                    # wp for partNet
-                    "2022preEE": 0.245, 
+                    # wp for particleNet
+                    "2022preEE": 0.245,
                     "2022postEE": 0.2605,
                     "2023preBPix": 0.1917,
                     "2023postBPix": 0.1919,
-                    "2024": 0.1272,
-                    "2025": 0.1272,
-                    "2026": 0.1272,
-                }
+                },
+                default=0.1272,  # 2024, 2025, 2026 UParT
             ),
             # jet puID
             "jet_puid": EraModifier(
@@ -301,23 +266,13 @@ def build_config(
                     "2016postVFP": 1,  # 0==fail, 1==pass(loose), 3==pass(loose,medium), 7==pass(loose,medium,tight)
                     "2017": 4,  # 0==fail, 4==pass(loose), 6==pass(loose,medium), 7==pass(loose,medium,tight)
                     "2018": 4,  # 0==fail, 4==pass(loose), 6==pass(loose,medium), 7==pass(loose,medium,tight)
-                    "2022preEE": 0, #not used
-                    "2022postEE": 0, #not used
-                    "2023preBPix": 0, #not used
-                    "2023postBPix": 0, #not used
-                    "2024": 0, #not used
-                    "2025": 0, #not used
-                    "2026": 0, #not used
-                }
+                },
+                default=0,  # not used from Run 3 onwards
             ),
             "jet_puid_max_pt": 50,  # recommended to apply puID only for jets below 50 GeV
-            # jet energy calibration 
+            # jet energy calibration
             "jet_id_json": EraModifier(
                 {
-                    "2016preVFP": '""',
-                    "2016postVFP": '""',
-                    "2017": '""',
-                    "2018": '""',
                     "2022preEE": "/cvmfs/cms-griddata.cern.ch/cat/metadata/JME/Run3-22CDSep23-Summer22-NanoAODv12/2026-06-05/jetid.json.gz",
                     "2022postEE": "/cvmfs/cms-griddata.cern.ch/cat/metadata/JME/Run3-22EFGSep23-Summer22EE-NanoAODv12/2026-06-05/jetid.json.gz",
                     "2023preBPix": "/cvmfs/cms-griddata.cern.ch/cat/metadata/JME/Run3-23CSep23-Summer23-NanoAODv12/2026-06-05/jetid.json.gz",
@@ -325,7 +280,8 @@ def build_config(
                     "2024": "/cvmfs/cms-griddata.cern.ch/cat/metadata/JME/Run3-24CDEReprocessingFGHIPrompt-Summer24-NanoAODv15/2026-07-14/jetid.json.gz",
                     "2025": "/cvmfs/cms-griddata.cern.ch/cat/metadata/JME/Run3-25Prompt-Summer24-NanoAODv15/2026-07-14/jetid.json.gz",
                     "2026": "/cvmfs/cms-griddata.cern.ch/cat/metadata/JME/Run3-26Prompt-Summer24-NanoAODv15/2026-07-15/jetid.json.gz",
-                }
+                },
+                default='""',  # not used for Run 2
             ),
             "jet_collection_name": "AK4PUPPI", #only used for jet ID so not relevant for run 2
             "jet_jec_file": EraModifier(
@@ -385,14 +341,8 @@ def build_config(
                     "2016postVFP": "AK4PFchs",
                     "2017": "AK4PFchs",
                     "2018": "AK4PFchs",
-                    "2022preEE": "AK4PFPuppi",
-                    "2022postEE": "AK4PFPuppi",
-                    "2023preBPix": "AK4PFPuppi",
-                    "2023postBPix": "AK4PFPuppi",
-                    "2024": "AK4PFPuppi",
-                    "2025": "AK4PFPuppi",
-                    "2026": "AK4PFPuppi",
-                }
+                },
+                default="AK4PFPuppi", # run3
             ),
             # jet veto configuration
             "jet_veto_map_file": EraModifier(
@@ -446,20 +396,17 @@ def build_config(
             {
                 "btag_eff_file": EraModifier(
                     {
-                        "2016preVFP": "TO_ADD",
-                        "2016postVFP": "TO_ADD",
-                        "2017": "TO_ADD",
-                        "2018": "TO_ADD",
                         "2022preEE": f"payloads/btag_efficiencies/2022preEE/{ch}/btag_efficiency.json.gz",
                         "2022postEE": f"payloads/btag_efficiencies/2022postEE/{ch}/btag_efficiency.json.gz",
                         "2023preBPix": f"payloads/btag_efficiencies/2023preBPix/{ch}/btag_efficiency.json.gz",
                         "2023postBPix": f"payloads/btag_efficiencies/2023postBPix/{ch}/btag_efficiency.json.gz",
                         "2024": f"payloads/btag_efficiencies/2024/{ch}/btag_efficiency.json.gz",
                         "2025": f"payloads/btag_efficiencies/2025/{ch}/btag_efficiency.json.gz",
-                        "2026": f"payloads/btag_efficiencies/2025/{ch}/btag_efficiency.json.gz", # to update later on
-                    }
+                        "2026": f"payloads/btag_efficiencies/2025/{ch}/btag_efficiency.json.gz",  # to update later on
+                    },
+                    default='""',  # Run 2
                 ),
-            }
+            },
         )
     configuration.add_config_parameters(
         scopes,
@@ -480,13 +427,7 @@ def build_config(
                     "2026": "/cvmfs/cms-griddata.cern.ch/cat/metadata/BTV/Run3-25Prompt-Summer24-NanoAODv15/2026-06-26/btagging.json.gz",
                 }
             ),
-            "btag_sf_variation": "central",  # used by BtaggingShape (Run2, deepJet_shape)
-            # Used by BtaggingWP (Run3, fixedWP SFs). The BTV POG measures
-            # b/c-flavor (mujets/comb) and light-flavor (incl) scale factors
-            # with independent methods and independent uncertainty sources
-            # (see https://btv-wiki.docs.cern.ch/ScaleFactors/#sf-uncertainties-and-correlations-across-years),
-            # so they need two independent variation knobs: shifting one must
-            # leave the other at "central".
+            "btag_sf_variation": "central",
             "btag_sf_variation_bc": "central",
             "btag_sf_variation_lf": "central",
             "btag_wp": "M",
@@ -605,19 +546,13 @@ def build_config(
             # pair selection
             "pairselection_min_dR": 0.5,
             # propagate jet and lepton sf correction to the met
-            "propagateLeptons": SampleModifier(
-                {"data": False},
-                default=True,
-            ),
+            "propagateLeptons": True,
             "propagateJets": True,
-            # propagate the (NanoAOD-derived) unclustered energy variation onto
-            # the analysis MET, see met.ApplyUnclusteredMetShift
-            "propagateUnclustered": True,
             # recoil corrections
             "recoil_corrections_file": EraModifier(
                 {
-                    "2016preVFP": "data/recoil_corrections/Type1_PuppiMET_2016.root",  # These are likely from Legacy data sets, therefore no difference in pre and postVFP
-                    "2016postVFP": "data/recoil_corrections/Type1_PuppiMET_2016.root",  # These are likely from Legacy data sets, therefore no difference in pre and postVFP
+                    "2016preVFP": "data/recoil_corrections/Type1_PuppiMET_2016.root",  
+                    "2016postVFP": "data/recoil_corrections/Type1_PuppiMET_2016.root",  
                     "2017": "data/recoil_corrections/Type1_PuppiMET_2017.root",
                     "2018": "data/recoil_corrections/Type1_PuppiMET_2018.root",
                     "2022preEE": "data/hleprare/RecoilCorrlib/Recoil_corrections_2022preEE_v5.json.gz",
@@ -631,8 +566,8 @@ def build_config(
             ),
             "recoil_systematics_file": EraModifier(
                 {
-                    "2016preVFP": "data/recoil_corrections/PuppiMETSys_2016.root",  # These are likely from Legacy data sets, therefore no difference in pre and postVFP
-                    "2016postVFP": "data/recoil_corrections/PuppiMETSys_2016.root",  # These are likely from Legacy data sets, therefore no difference in pre and postVFP
+                    "2016preVFP": "data/recoil_corrections/PuppiMETSys_2016.root",  
+                    "2016postVFP": "data/recoil_corrections/PuppiMETSys_2016.root",  
                     "2017": "data/recoil_corrections/PuppiMETSys_2017.root",
                     "2018": "data/recoil_corrections/PuppiMETSys_2018.root",
                     "2022preEE": '""',
@@ -646,16 +581,6 @@ def build_config(
             ),
             "recoil_method": "QuantileMapHist", #other option is pure "Resclaing"
             "recoil_variation": "nom",
-            # Recoil response/resolution uncertainty (Run3, correctionlib-based).
-            # Per the HLepRare recoil documentation, these uncertainties must be
-            # evaluated on top of the nominally (QuantileMapHist) recoil-corrected
-            # MET, not on the pre-recoil-correction MET. This is a separate
-            # producer (met.ApplyRecoilUncertainty) chained after the nominal
-            # recoil correction; applyRecoilUncertainty defaults to False (no-op)
-            # and is only switched on for the metRecoilResponse/metRecoilResolution
-            # shifts, see below.
-            "applyRecoilUncertainty": False,
-            "recoil_uncertainty_variation": "RespUp",
             "applyRecoilCorrections": SampleModifier( #apply only to single boson processes
                 {
                     "dyjets": True,
@@ -684,8 +609,8 @@ def build_config(
             # zpt reweighting for DY samples
             "zpt_file": EraModifier(
                 {
-                    "2016preVFP": "data/zpt/htt_scalefactors_legacy_2016.root",  # ToDO: Measured in legacy, therefore the same for pre- and postVFP for now
-                    "2016postVFP": "data/zpt/htt_scalefactors_legacy_2016.root",  # ToDO: Measured in legacy, therefore the same for pre- and postVFP for now
+                    "2016preVFP": "data/zpt/htt_scalefactors_legacy_2016.root",  
+                    "2016postVFP": "data/zpt/htt_scalefactors_legacy_2016.root",  
                     "2017": "data/zpt/htt_scalefactors_legacy_2017.root",
                     "2018": "data/zpt/htt_scalefactors_legacy_2018.root",
                     "2022preEE": "data/hleprare/DYweightCorrlib/DY_pTll_weights_2022preEE_v5.json.gz",
@@ -693,12 +618,6 @@ def build_config(
                     "2023preBPix": "data/hleprare/DYweightCorrlib/DY_pTll_weights_2023preBPix_v5.json.gz",
                     "2023postBPix": "data/hleprare/DYweightCorrlib/DY_pTll_weights_2023postBPix_v5.json.gz",
                     "2024": "data/hleprare/DYweightCorrlib/DY_pTll_weights_2024_v5.json.gz",
-                    # NOTE: as of the v5 correctionlib recommendation
-                    # (https://cms-higgs-leprare.docs.cern.ch/htt-common/DY_reweight/),
-                    # no dedicated DY pTll reweighting has been derived yet for
-                    # 2025/2026 data-taking. The 2024 corrections are reused as a
-                    # placeholder until a dedicated recommendation is published;
-                    # revisit once 2025/2026 corrections become available.
                     "2025": "data/hleprare/DYweightCorrlib/DY_pTll_weights_2024_v5.json.gz",
                     "2026": "data/hleprare/DYweightCorrlib/DY_pTll_weights_2024_v5.json.gz",
                 }
@@ -717,7 +636,7 @@ def build_config(
                     "wjets": "NNLO",
                     }, 
                 default="NLO",
-            ), #from GrASP it looks like the DY powheg samples are also NLO and not NNLO
+            ), #from GrASP it looks like the DY powheg samples are also NLO and not NNLO, doesn't matter for the corrections
             "zpt_variation": "nom",
 
             # STXS weights
@@ -822,14 +741,7 @@ def build_config(
                     "2023preBPix": "/cvmfs/cms-griddata.cern.ch/cat/metadata/TAU/Run3-23CSep23-Summer23-NanoAODv12/2025-12-25/tau.json.gz",
                     "2023postBPix": "/cvmfs/cms-griddata.cern.ch/cat/metadata/TAU/Run3-23DSep23-Summer23BPix-NanoAODv12/2025-12-25/tau.json.gz",
                     "2024": "/cvmfs/cms-griddata.cern.ch/cat/metadata/TAU/Run3-24CDEReprocessingFGHIPrompt-Summer24-NanoAODv15/2026-01-14/tau.json.gz",
-                    # 2025: TauPOG explicitly recommends reusing the 2024 (Summer24-NanoAODv15)
-                    # corrections for 2025, since the same MC campaign is used
-                    # (https://tau-wiki.docs.cern.ch, "NanoAOD version" section).
                     "2025": "/cvmfs/cms-griddata.cern.ch/cat/metadata/TAU/Run3-24CDEReprocessingFGHIPrompt-Summer24-NanoAODv15/2026-01-14/tau.json.gz",
-                    # 2026: NOT an official recommendation yet -- as of the TauPOG Wiki
-                    # (2026-06-10) all 2026 tau corrections (ID SF, ES, e/mu fake SF, trigger)
-                    # are marked unavailable ("Not available yet"). This reuses the 2024/2025
-                    # json as a placeholder; replace once TauPOG publishes 2026 corrections.
                     "2026": "/cvmfs/cms-griddata.cern.ch/cat/metadata/TAU/Run3-24CDEReprocessingFGHIPrompt-Summer24-NanoAODv15/2026-01-14/tau.json.gz",
                 }
             ),
@@ -930,8 +842,8 @@ def build_config(
             ),
             "muon_id_sf_name": "NUM_MediumID_DEN_TrackerMuons",  # correction for mediumId WP
             "muon_iso_sf_name": "NUM_TightPFIso_DEN_MediumID",  # correction for TightPFIso WP (PF isolation < 0.15)
-            "muon_id_variation": "nominal",  # "systup"/"systdown" are up/down variations
-            "muon_iso_variation": "nominal",  # "systup"/"systdown" are up/down variations
+            "muon_id_variation": "nominal",  # "systup"/"systdown/statup/statdown", user defined
+            "muon_iso_variation": "nominal",  # "systup"/"systdown/statup/statdown", user defined
 
             #run 2 scale factors from embedding framework
             "mc_muon_sf_file": EraModifier(
@@ -1217,66 +1129,23 @@ def build_config(
                 "tau_id_vsjet_3prong1pizero": "nom",
             }
         )
-        configuration.add_config_parameters(
-            ["mt", "mm", "em"],
-            {
-                "muon_iso_sf_name": "NUM_TightRelIso_DEN_MediumID",
-            }
-        )
-        configuration.add_config_parameters(
-            ["et", "ee", "em"],
-            {
-                "ele_id_sf_name": "UL-Electron-ID-SF",
-            }
-        )
-        configuration.add_config_parameters(
-            ["et", "mt"],
-            {
-                "vsjet_tau_id_bit": 1,
-                "vsele_tau_id_bit": 1,
-                "vsmu_tau_id_bit": 1,
-            }
-        )
-        configuration.add_config_parameters(
-            ["mt"],
-            {
-                "max_muon_eta": 2.1,
-                "max_muon_iso": 0.3,
-            }
-        )
-        configuration.add_config_parameters(
-            ["em", "ee"],
-            {
-                "max_ele_eta": 2.1,
-                "max_ele_iso": 0.3,
-                "max_muon_eta": 2.1,
-            },
-        )
-        configuration.add_config_parameters(
-            ["mm"],
-            {
-                "min_muon_pt": 20.0,
-                "max_muon_eta": 2.1,
-                "max_muon_iso": 0.15,
-            }
-        )
-        configuration.add_config_parameters(
-            ["et"],
-            {
-                "max_electron_eta": 2.1,
-                "max_ele_iso": 0.5,
-            }
-        )
-        configuration.add_config_parameters(
-            ["tt"],
-            {
+        for chs, params in [
+            (["mt", "mm", "em"], {"muon_iso_sf_name": "NUM_TightRelIso_DEN_MediumID"}),
+            (["et", "ee", "em"], {"ele_id_sf_name": "UL-Electron-ID-SF"}),
+            (["et", "mt"], {"vsjet_tau_id_bit": 1, "vsele_tau_id_bit": 1, "vsmu_tau_id_bit": 1}),
+            (["mt"], {"max_muon_eta": 2.1, "max_muon_iso": 0.3}),
+            (["em", "ee"], {"max_ele_eta": 2.1, "max_ele_iso": 0.3, "max_muon_eta": 2.1}),
+            (["mm"], {"min_muon_pt": 20.0, "max_muon_eta": 2.1, "max_muon_iso": 0.15}),
+            (["et"], {"max_electron_eta": 2.1, "max_ele_iso": 0.5}),
+            (["tt"], {
                 "tau_vsjet_sf_dependence": "pt",
                 "min_tau_pt": 35.0,
                 "vsjet_tau_id_bit": 4,
                 "vsele_tau_id_bit": 4,
                 "vsmu_tau_id_bit": 1,
-            }
-        )
+            }),
+        ]:
+            configuration.add_config_parameters(chs, params)
 
     ############################
     ######## Producers #########
@@ -1284,7 +1153,6 @@ def build_config(
     configuration.add_producers(
         "global",
         [
-            # event.RunLumiEventFilter,
             event.SampleFlags,
             event.Lumi,
             event.npartons,
@@ -1461,7 +1329,6 @@ def build_config(
             triggers.EMGenerateSingleMuonTriggerFlags,
             scalefactors.SingleEleTriggerSF,
             scalefactors.SingleMuTriggerSF,
-            #triggers.EMGenerateCrossTriggerFlags,
         ],
     )
     configuration.add_producers(
@@ -1483,8 +1350,6 @@ def build_config(
             genparticles.TTGenDiTauPairQuantities,
             scalefactors.TauID_SF,
             triggers.TTGenerateDoubleTauTriggerFlags,
-            #triggers.GenerateSingleTrailingTauTriggerFlags,
-            #triggers.GenerateSingleLeadingTauTriggerFlags,
             scalefactors.DoubleTauTriggerSF,
         ],
     )
@@ -1492,570 +1357,111 @@ def build_config(
     ################################
     ######### Modifications ########
     ################################
-    
-    if era == "2024":
-        # separate MC for 2024 and 2025 by even/odd event number
-        configuration.add_modification_rule(
-            "global",
-            AppendProducer(
-                producers=[event.EvenIDFilter],
-                exclude_samples=["data", "embedding"],
-            ),
-        )
-    if era == "2025" or era == "2026":
-        # temporary root pileup for data 2025 by tau fw group, 23/03/2026
-        # configuration.add_modification_rule(
-        #     "global",
-        #     ReplaceProducer(
-        #         producers=[event.PUweights, event.PUweights_root],
-        #         exclude_samples=["data", "embedding", "embedding_mc"],
-        #     ),
-        # )
-        # separate MC for 2024 and 2025 by even/odd event number
-        configuration.add_modification_rule(
-            "global",
-            AppendProducer(
-                producers=[event.OddIDFilter],
-                exclude_samples=["data", "embedding"],
-            ),
-        )
 
-    configuration.add_modification_rule(
-        "global",
-        RemoveProducer(
-            producers=[event.npartons],
-            exclude_samples=["dyjets", "dyjets_powheg", "dyjets_amcatnlo", "dyjets_amcatnlo_ll", "dyjets_amcatnlo_tt", "wjets", "wjets_amcatnlo", "electroweak_boson"],
-        ),
-    )
-    configuration.add_modification_rule(
-        "global",
-        RemoveProducer(
-            producers=[
-                event.PUweights,
-                event.PS_weight,
-                ],
-            samples=["data", "embedding", "embedding_mc"],
-        ),
-    )
-    configuration.add_modification_rule(
-        "global",
-        RemoveProducer(
-            producers=[
-                event.LHE_Scale_weight,
-                event.LHE_PDF_weight,
-                event.LHE_alphaS_weight,
-                ],
-            samples=["data", "embedding", "embedding_mc", "diboson"],
-        ),
-    )
-    configuration.add_modification_rule(
-        "global",
-        AppendProducer(
-            producers=event.JSONFilter, 
-            samples=["data", "embedding", "embedding_mc"]),
-    )
-    ## producer to add a cut on DYto2L affected by pythia bug where DYto2Tau has been reprocessed
-    configuration.add_modification_rule(
-        "global",
-        AppendProducer(
-            producers=[genparticles.GenDYFlavor,genparticles.GenDYFilter],
-            samples=["dyjets_amcatnlo_ll"],
-        ),
-    )
-    configuration.add_modification_rule(
-        scopes,
-        RemoveProducer(
-            producers=[genparticles.GenMatching],
-            samples=["data", "embedding", "embedding_mc"],
-        ),
-    )
-    configuration.add_modification_rule(
-        "global",
-        ReplaceProducer(
-            producers=[jets.GenJet, jets.GenJet_data],
-            samples=["data", "embedding", "embedding_mc"],
-        ),
-    )
-    configuration.add_modification_rule(
-        scopes,
-        RemoveProducer(
-            producers=[scalefactors.btaggingWP_SF],
-            samples=["data", "embedding", "embedding_mc"],
-        ),
-    )
-    configuration.add_modification_rule(
-        scopes,
-        AppendProducer(
-            producers=event.ZPtReweighting,
-            samples=["dyjets_powheg", "dyjets_amcatnlo", "dyjets_amcatnlo_ll", "dyjets_amcatnlo_tt", "electroweak_boson"]
-        ),
-    )
-    configuration.add_modification_rule(
-        scopes,
-        AppendProducer(
-            producers=[event.GGH_NNLO_Reweighting, event.GGH_WG1_Uncertainties],
-            samples=["ggh_htautau", "rem_htautau"],
-        ),
-    )
-    configuration.add_modification_rule(
-        scopes,
-        AppendProducer(
-            producers=event.QQH_WG1_Uncertainties,
-            samples=["vbf_htautau", "rem_htautau"],
-        ),
-    )
-    configuration.add_modification_rule(
-        scopes,
-        AppendProducer(
-            producers=[event.TopPtReweighting], 
-            samples=["ttbar",]
-        ),
-    )
-    configuration.add_modification_rule(
-        ["et", "mt", "tt"],
-        ReplaceProducer(
-            producers=[taus.TauEnergyCorrection, taus.TauEnergyCorrection_data],
-            samples=["data", "embedding", "embedding_mc"],
-        ),
-    )
-    configuration.add_modification_rule(
-        ["et", "mt", "tt"],
-        RemoveProducer(
-            producers=[
-                scalefactors.TauID_SF,
-            ],
-            samples=["data", "embedding", "embedding_mc"],
-        ),
-    )
-    configuration.add_modification_rule(
-        ["mt", "em", "mm"],
-        RemoveProducer(
-            producers=[
-                scalefactors.MuonIDIso_SF,
-            ],
-            samples=["data", "embedding", "embedding_mc"],
-        ),
-    )
-    configuration.add_modification_rule(
-        ["et", "ee", "em"],
-        RemoveProducer(
-            producers=[
-                scalefactors.EleID_SF,
-            ],
-            samples=["data", "embedding", "embedding_mc"],
-        ),
-    )
-    configuration.add_modification_rule(
-        ["mt"],
-        RemoveProducer(
-            producers=[genparticles.MTGenDiTauPairQuantities],
-            samples=["data"],
-        ),
-    )
-    configuration.add_modification_rule(
-        ["mm"],
-        RemoveProducer(
-            producers=[genparticles.MuMuGenPairQuantities],
-            samples=["data"],
-        ),
-    )
-    configuration.add_modification_rule(
-        ["et"],
-        RemoveProducer(
-            producers=[genparticles.ETGenDiTauPairQuantities],
-            samples=["data"],
-        ),
-    )
-    configuration.add_modification_rule(
-        ["em"],
-        RemoveProducer(
-            producers=[genparticles.EMGenDiTauPairQuantities],
-            samples=["data"],
-        ),
-    )
-    configuration.add_modification_rule(
-        ["ee"],
-        RemoveProducer(
-            producers=[genparticles.ElElGenPairQuantities],
-            samples=["data"],
-        ),
-    )
-    configuration.add_modification_rule(
-        ["tt"],
-        RemoveProducer(
-            producers=[
-                genparticles.TTGenDiTauPairQuantities
-            ],
-            samples=["data"],
-        ),
-    ) 
+    MC_ONLY = ["data", "embedding", "embedding_mc"]
+    
+    for mod_scopes, rule_cls, producers, sample_filter in [
+        ("global", RemoveProducer, [event.npartons], 
+            {"exclude_samples": ["dyjets", "dyjets_powheg", "dyjets_amcatnlo", "dyjets_amcatnlo_ll", "dyjets_amcatnlo_tt", "wjets", "wjets_amcatnlo", "electroweak_boson"]}),
+        ("global", RemoveProducer, [event.PUweights, event.PS_weight], {"samples": MC_ONLY}),
+        ("global", RemoveProducer, [event.LHE_Scale_weight, event.LHE_PDF_weight, event.LHE_alphaS_weight], {"samples": MC_ONLY + ["diboson"]}),
+        (scopes, RemoveProducer, [scalefactors.btaggingWP_SF], {"samples": MC_ONLY}),
+        (scopes, RemoveProducer, [genparticles.GenMatching], {"samples": MC_ONLY}),
+        (["et", "mt", "tt"], RemoveProducer, [scalefactors.TauID_SF], {"samples": MC_ONLY}),
+        (["mt", "em", "mm"], RemoveProducer, [scalefactors.MuonIDIso_SF], {"samples": MC_ONLY}),
+        (["et", "ee", "em"], RemoveProducer, [scalefactors.EleID_SF], {"samples": MC_ONLY}),
+        (["mt"], RemoveProducer, [genparticles.MTGenDiTauPairQuantities], {"samples": ["data"]}),
+        (["mm"], RemoveProducer, [genparticles.MuMuGenPairQuantities], {"samples": ["data"]}),
+        (["et"], RemoveProducer, [genparticles.ETGenDiTauPairQuantities], {"samples": ["data"]}),
+        (["em"], RemoveProducer, [genparticles.EMGenDiTauPairQuantities], {"samples": ["data"]}),
+        (["ee"], RemoveProducer, [genparticles.ElElGenPairQuantities], {"samples": ["data"]}),
+        (["tt"], RemoveProducer, [genparticles.TTGenDiTauPairQuantities], {"samples": ["data"]}),
+
+        ("global", AppendProducer, [event.JSONFilter], {"samples": MC_ONLY}),
+        ## producer to add a cut on DYto2L affected by pythia bug where DYto2Tau has been reprocessed
+        ("global", AppendProducer, [genparticles.GenDYFlavor, genparticles.GenDYFilter], {"samples": ["dyjets_amcatnlo_ll"]}),
+        (scopes, AppendProducer, [event.ZPtReweighting], {"samples": ["dyjets_powheg", "dyjets_amcatnlo", "dyjets_amcatnlo_ll", "dyjets_amcatnlo_tt", "electroweak_boson"]}),
+        (scopes, AppendProducer, [event.GGH_NNLO_Reweighting, event.GGH_WG1_Uncertainties], {"samples": ["ggh_htautau", "rem_htautau"]}),
+        (scopes, AppendProducer, [event.QQH_WG1_Uncertainties], {"samples": ["vbf_htautau", "rem_htautau"]}),
+        (scopes, AppendProducer, [event.TopPtReweighting], {"samples": ["ttbar"]}),
+                
+        ("global", ReplaceProducer, [jets.GenJet, jets.GenJet_data], {"samples": MC_ONLY}),
+        (["et", "mt", "tt"], ReplaceProducer, [taus.TauEnergyCorrection, taus.TauEnergyCorrection_data], {"samples": MC_ONLY}),
+        
+    ]:
+        configuration.add_modification_rule(mod_scopes, rule_cls(producers=producers, **sample_filter))
 
     if int(era[:4]) < 2022:
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers = [electrons.ElectronIDCut, electrons.ElectronIDCut_v9],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers = [event.TopPtReweighting, event.TopPtReweighting_Run2],
-                exclude_samples=["ttbar"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[electrons.ElectronPtCorrectionMC, electrons.ElectronPtCorrectionMC_v9],
-                exclude_samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[electrons.ElectronPtCorrectionMC, electrons.RenameElectronPt],
-                samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[event.DiLeptonVeto, event.DiLeptonVeto_v9],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[jets.JetEnergyCorrection, jets.JetEnergyCorrection_v12],
-                exclude_samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[jets.JetEnergyCorrection, jets.JetEnergyCorrection_data],
-                samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[jets.JetID, jets.JetID_rename],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[jets.JetBTagUParT, jets.JetBTagDeep],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[jets.JetRho, jets.JetRho_v9],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[jets.GoodJets, jets.GoodJets_Run2],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[jets.GoodBJets, jets.GoodBJets_Run2],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            RemoveProducer(
-                producers=[jets.JetVetoMapVeto],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            scopes,
-            ReplaceProducer(
-                producers=[scalefactors.btaggingWP_SF,scalefactors.btagging_SF],
-                exclude_samples=["data", "embedding", "embedding_mc"], 
-            ),
-        )
-        configuration.add_modification_rule(
-            scopes,
-            ReplaceProducer(
-                producers=[met.MetCorrections, met.MetCorrections_Run2],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            scopes,
-            ReplaceProducer(
-                producers=[met.PFMetCorrections, met.PFMetCorrections_Run2],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["mt", "et", "tt"],
-            ReplaceProducer(
-                producers=[taus.BaseTaus, taus.BaseTaus_v9],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["mt", "et", "tt"],
-            ReplaceProducer(
-                producers=[taus.GoodTaus, taus.GoodTaus_v9],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["mt", "et", "tt"],
-            ReplaceProducer(
-                producers=[scalefactors.TauID_SF, scalefactors.TauID_SF_v9],
-                exclude_samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["et", "mt", "tt"],
-            ReplaceProducer(
-                producers=[taus.TauEnergyCorrection, taus.TauEnergyCorrection_ES_dm_pt_binned],
-                exclude_samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["tt"],
-            ReplaceProducer(
-                producers=[pairquantities.TTDiTauPairQuantities, pairquantities.TTDiTauPairQuantities_v9],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["mt"],
-            ReplaceProducer(
-                producers=[pairquantities.MTDiTauPairQuantities, pairquantities.MTDiTauPairQuantities_v9],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["et"],
-            ReplaceProducer(
-                producers=[pairquantities.ETDiTauPairQuantities, pairquantities.ETDiTauPairQuantities_v9],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["mt", "em", "mm"],
-            RemoveProducer(
-                producers=[
-                    scalefactors.MuonIDIso_SF,
-                ],
-                exclude_samples=["data", "embedding", "embedding_mc"], 
-            ),
-        )
-        configuration.add_modification_rule(
-            ["et", "ee", "em"],
-            RemoveProducer(
-                producers=[
-                    scalefactors.EleID_SF,
-                ],
-                exclude_samples=["data", "embedding", "embedding_mc"], 
-            ),
-        )
-        configuration.add_modification_rule(
-            ["mt"],
-            AppendProducer(
-                producers=[
-                    triggers.MTGenerateCrossTriggerFlags,
-                    triggers.GenerateSingleTrailingTauTriggerFlags
-                    ],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["mt"],
-            AppendProducer(
-                producers=[
-                    scalefactors.MTGenerateSingleMuonTriggerSF_MC,
-                    scalefactors.PrivateMuonIDSF_1_MC,
-                    scalefactors.PrivateMuonIsoSF_1_MC,
-                ],
-                exclude_samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["mt"],
-            RemoveProducer(
-                producers=[
-                    scalefactors.SingleMuTriggerSF,
-                ],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["et"],
-            AppendProducer(
-                producers=[
-                    triggers.ETGenerateCrossTriggerFlags,
-                    triggers.GenerateSingleTrailingTauTriggerFlags],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["et"],
-            AppendProducer(
-                producers=[
-                    scalefactors.ETGenerateSingleElectronTriggerSF_MC,
-                    scalefactors.PrivateElectronIDSF_1_MC,
-                    scalefactors.PrivateElectronIsoSF_1_MC,
-                ],
-                exclude_samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["et"],
-            RemoveProducer(
-                producers=[
-                    scalefactors.SingleEleTriggerSF,
-                ],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["tt"],
-            AppendProducer(
-                producers=[
-                    triggers.GenerateSingleTrailingTauTriggerFlags,
-                    triggers.GenerateSingleLeadingTauTriggerFlags,
-                    ],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["tt"],
-            RemoveProducer(
-                producers=[
-                    scalefactors.DoubleTauTriggerSF,
-                ],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["em"],
-            AppendProducer(
-                producers=[
-                    scalefactors.PrivateElectronIDSF_1_MC,
-                    scalefactors.PrivateElectronIsoSF_1_MC,
-                    scalefactors.PrivateMuonIDSF_2_MC,
-                    scalefactors.PrivateMuonIsoSF_2_MC,
-                ],
-                exclude_samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["mm"],
-            AppendProducer(
-                producers=[
-                    scalefactors.PrivateMuonIDSF_1_MC,
-                    scalefactors.PrivateMuonIsoSF_1_MC,
-                    scalefactors.PrivateMuonIDSF_2_MC,
-                    scalefactors.PrivateMuonIsoSF_2_MC,
-                    scalefactors.MTGenerateSingleMuonTriggerSF_MC,
-                ],
-                exclude_samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["ee"],
-            AppendProducer(
-                producers=[
-                    scalefactors.PrivateElectronIDSF_1_MC,
-                    scalefactors.PrivateElectronIsoSF_1_MC,
-                    scalefactors.PrivateElectronIDSF_2_MC,
-                    scalefactors.PrivateElectronIsoSF_2_MC,
-                    scalefactors.ETGenerateSingleElectronTriggerSF_MC,
-                ],
-                exclude_samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
+        for mod_scopes, rule_cls, producers, sample_filter in [
+            ("global", RemoveProducer, [jets.JetVetoMapVeto], {"exclude_samples": ["fake_era"]}),
+            (["mt", "em", "mm"], RemoveProducer, [scalefactors.MuonIDIso_SF], {"exclude_samples": MC_ONLY}),
+            (["et", "ee", "em"], RemoveProducer, [scalefactors.EleID_SF], {"exclude_samples": MC_ONLY}),
+            (["mt"], RemoveProducer, [scalefactors.SingleMuTriggerSF], {"exclude_samples": ["fake_era"]}),
+            (["et"], RemoveProducer, [scalefactors.SingleEleTriggerSF], {"exclude_samples": ["fake_era"]}),
+            (["tt"], RemoveProducer, [scalefactors.DoubleTauTriggerSF], {"exclude_samples": ["fake_era"]}),
+
+            # cross triggers and embedding triggers      
+            (["mt"], AppendProducer, [triggers.MTGenerateCrossTriggerFlags, triggers.GenerateSingleTrailingTauTriggerFlags], {"exclude_samples": ["fake_era"]}),
+            (["mt"], AppendProducer, [scalefactors.MTGenerateSingleMuonTriggerSF_MC, scalefactors.PrivateMuonIDSF_1_MC, scalefactors.PrivateMuonIsoSF_1_MC], {"exclude_samples": MC_ONLY}),
+            (["et"], AppendProducer, [triggers.ETGenerateCrossTriggerFlags, triggers.GenerateSingleTrailingTauTriggerFlags], {"exclude_samples": ["fake_era"]}),
+            (["et"], AppendProducer, [scalefactors.ETGenerateSingleElectronTriggerSF_MC, scalefactors.PrivateElectronIDSF_1_MC, scalefactors.PrivateElectronIsoSF_1_MC], {"exclude_samples": MC_ONLY}),
+            (["tt"], AppendProducer, [triggers.GenerateSingleTrailingTauTriggerFlags, triggers.GenerateSingleLeadingTauTriggerFlags], {"exclude_samples": ["fake_era"]}),
+            (["em"], AppendProducer, [scalefactors.PrivateElectronIDSF_1_MC, scalefactors.PrivateElectronIsoSF_1_MC, scalefactors.PrivateMuonIDSF_2_MC, scalefactors.PrivateMuonIsoSF_2_MC], {"exclude_samples": MC_ONLY}),
+            (["mm"], AppendProducer, [scalefactors.PrivateMuonIDSF_1_MC, scalefactors.PrivateMuonIsoSF_1_MC, scalefactors.PrivateMuonIDSF_2_MC, scalefactors.PrivateMuonIsoSF_2_MC, scalefactors.MTGenerateSingleMuonTriggerSF_MC], {"exclude_samples": MC_ONLY}),
+            (["ee"], AppendProducer, [scalefactors.PrivateElectronIDSF_1_MC, scalefactors.PrivateElectronIsoSF_1_MC, scalefactors.PrivateElectronIDSF_2_MC, scalefactors.PrivateElectronIsoSF_2_MC, scalefactors.ETGenerateSingleElectronTriggerSF_MC], {"exclude_samples": MC_ONLY}),
+                        
+            ("global", ReplaceProducer, [electrons.ElectronIDCut, electrons.ElectronIDCut_v9], {"exclude_samples": ["fake_era"]}),
+            ("global", ReplaceProducer, [event.TopPtReweighting, event.TopPtReweighting_Run2], {"samples": ["ttbar"]}),
+            ("global", ReplaceProducer, [electrons.ElectronPtCorrectionMC, electrons.ElectronPtCorrectionMC_v9], {"exclude_samples": MC_ONLY}),
+            ("global", ReplaceProducer, [electrons.ElectronPtCorrectionMC, electrons.RenameElectronPt], {"samples": MC_ONLY}),
+            ("global", ReplaceProducer, [event.DiLeptonVeto, event.DiLeptonVeto_v9], {"exclude_samples": ["fake_era"]}),
+            ("global", ReplaceProducer, [jets.JetEnergyCorrection, jets.JetEnergyCorrection_v12], {"exclude_samples": MC_ONLY}),
+            ("global", ReplaceProducer, [jets.JetEnergyCorrection, jets.JetEnergyCorrection_data], {"samples": MC_ONLY}),
+            ("global", ReplaceProducer, [jets.JetID, jets.JetID_rename], {"exclude_samples": ["fake_era"]}),
+            ("global", ReplaceProducer, [jets.JetBTagUParT, jets.JetBTagDeep], {"exclude_samples": ["fake_era"]}),
+            ("global", ReplaceProducer, [jets.JetRho, jets.JetRho_v9], {"exclude_samples": ["fake_era"]}),
+            ("global", ReplaceProducer, [jets.GoodJets, jets.GoodJets_Run2], {"exclude_samples": ["fake_era"]}),
+            ("global", ReplaceProducer, [jets.GoodBJets, jets.GoodBJets_Run2], {"exclude_samples": ["fake_era"]}),
+            (scopes, ReplaceProducer, [scalefactors.btaggingWP_SF, scalefactors.btagging_SF], {"exclude_samples": MC_ONLY}),
+            (scopes, ReplaceProducer, [met.MetCorrections, met.MetCorrections_Run2], {"exclude_samples": ["fake_era"]}),
+            (scopes, ReplaceProducer, [met.PFMetCorrections, met.PFMetCorrections_Run2], {"exclude_samples": ["fake_era"]}),
+            (["mt", "et", "tt"], ReplaceProducer, [taus.BaseTaus, taus.BaseTaus_v9], {"exclude_samples": ["fake_era"]}),
+            (["mt", "et", "tt"], ReplaceProducer, [taus.GoodTaus, taus.GoodTaus_v9], {"exclude_samples": ["fake_era"]}),
+            (["mt", "et", "tt"], ReplaceProducer, [scalefactors.TauID_SF, scalefactors.TauID_SF_v9], {"exclude_samples": MC_ONLY}),
+            (["et", "mt", "tt"], ReplaceProducer, [taus.TauEnergyCorrection, taus.TauEnergyCorrection_ES_dm_pt_binned], {"exclude_samples": MC_ONLY}),
+            (["tt"], ReplaceProducer, [pairquantities.TTDiTauPairQuantities, pairquantities.TTDiTauPairQuantities_v9], {"exclude_samples": ["fake_era"]}),
+            (["mt"], ReplaceProducer, [pairquantities.MTDiTauPairQuantities, pairquantities.MTDiTauPairQuantities_v9], {"exclude_samples": ["fake_era"]}),
+            (["et"], ReplaceProducer, [pairquantities.ETDiTauPairQuantities, pairquantities.ETDiTauPairQuantities_v9], {"exclude_samples": ["fake_era"]}),
+        ]:
+            configuration.add_modification_rule(mod_scopes, rule_cls(producers=producers, **sample_filter))
+
         if era != "2018":
-            configuration.add_modification_rule(
-                "global",
-                AppendProducer(
-                    producers=event.PrefireWeight,
-                    exclude_samples=["fake_era"],
-                ),
-            )
-         # Broken sfs file for 2016. If nlo is used, this reweighting is not even needed. !!!
+            configuration.add_modification_rule("global", AppendProducer(producers=event.PrefireWeight, exclude_samples=["fake_era"],),)
+        # Broken sfs file for 2016. If nlo is used, this reweighting is not even needed. !!!
         if "2016" not in era:
-            configuration.add_modification_rule(
-                scopes,
-                AppendProducer(
-                    producers=event.ZPtReweighting_Run2, samples=["dyjets", "electroweak_boson"]
-                ),
-            )
+            configuration.add_modification_rule(scopes, AppendProducer(producers=event.ZPtReweighting_Run2, samples=["dyjets", "electroweak_boson"]),)
     else:
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[electrons.ElectronPtCorrectionMC, electrons.ElectronPtCorrectionData],
-                samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
+        configuration.add_modification_rule("global", ReplaceProducer(producers=[electrons.ElectronPtCorrectionMC, electrons.ElectronPtCorrectionData], samples=MC_ONLY,),)
 
     if 2022 <= int(era[:4]) < 2024:
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[jets.JetBTagUParT, jets.JetBTagPNet],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[jets.JetID, jets.JetIDRun3NanoV12Corrected],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            scopes,
-            ReplaceProducer(
-                producers=[met.MetCorrections, met.MetCorrections_v12],
-                exclude_samples=["fake_era"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["et", "mt", "tt"],
-            ReplaceProducer(
-                producers=[taus.TauEnergyCorrection, taus.TauEnergyCorrection_v12],
-                exclude_samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
-        configuration.add_modification_rule(
-            ["mt", "et", "tt"],
-            ReplaceProducer(
-                producers=[scalefactors.TauID_SF, scalefactors.TauID_SF_v12],
-                exclude_samples=["data", "embedding", "embedding_mc"],
-            ),
-        )
+        for mod_scopes, rule_cls, producers, sample_filter in [
+            ("global", ReplaceProducer, [jets.JetBTagUParT, jets.JetBTagPNet], {"exclude_samples": ["fake_era"]}),
+            ("global", ReplaceProducer, [jets.JetID, jets.JetIDRun3NanoV12Corrected], {"exclude_samples": ["fake_era"]}),
+            (scopes, ReplaceProducer, [met.MetCorrections, met.MetCorrections_v12], {"exclude_samples": ["fake_era"]}),
+            (["et", "mt", "tt"], ReplaceProducer, [taus.TauEnergyCorrection, taus.TauEnergyCorrection_v12], {"exclude_samples": MC_ONLY}),
+            (["mt", "et", "tt"], ReplaceProducer, [scalefactors.TauID_SF, scalefactors.TauID_SF_v12], {"exclude_samples": MC_ONLY}),
+        ]:
+            configuration.add_modification_rule(mod_scopes, rule_cls(producers=producers, **sample_filter))
     
     if int(era[:4]) < 2024:
-        configuration.add_modification_rule(
-            "global",
-            ReplaceProducer(
-                producers=[met.MetBasics, met.MetBasics_v12],
-                exclude_samples=["fake_era"],
-            ),
-        )
+        configuration.add_modification_rule("global", ReplaceProducer(producers=[met.MetBasics, met.MetBasics_v12], exclude_samples=["fake_era"],),)
+
+    # separate MC for 2024 and 2025 by even/odd event number
+    if era == "2024":
+        configuration.add_modification_rule("global", AppendProducer(producers=[event.EvenIDFilter], exclude_samples=MC_ONLY,),)
+    if era == "2025" or era == "2026":
+        configuration.add_modification_rule("global", AppendProducer(producers=[event.OddIDFilter], exclude_samples=["data", "embedding"],),)
+    
 
     #########################
     ######## OUTPUTS ########
@@ -2119,7 +1525,8 @@ def build_config(
             q.dimuon_veto,
             q.extraelec_veto,
             ] + [p for p in genparticles.MTGenDiTauPairQuantities.get_outputs("mt")
-            ],
+            ] + [p for p in scalefactors.TauID_SF.get_outputs("mt")
+        ],
     )
     configuration.add_outputs(
         "mm",
@@ -2140,7 +1547,8 @@ def build_config(
             q.dimuon_veto,
             q.extraelec_veto,
             ] + [p for p in genparticles.ETGenDiTauPairQuantities.get_outputs("et")
-            ],
+            ] + [p for p in scalefactors.TauID_SF.get_outputs("et")
+        ],
     )
     configuration.add_outputs(
         "em",
@@ -2179,22 +1587,21 @@ def build_config(
             q.dimuon_veto,
             q.extraelec_veto,
             ] + [p for p in genparticles.TTGenDiTauPairQuantities.get_outputs("tt")
-            ],
+            ] + [p for p in scalefactors.TauID_SF.get_outputs("tt")
+        ],
     )
 
+    if int(era[:4]) < 2024:
+        configuration.add_outputs("global", [p for p in met.MetBasics_v12.get_outputs("global")],)
+    else:
+        configuration.add_outputs("global", [p for p in met.MetBasics.get_outputs("global")],)
+
     if int(era[:4]) < 2022:
-        configuration.add_outputs(
-            "global",
-            [
-                p for p in met.MetBasics.get_outputs("global")
-            ],
-        )
         configuration.add_outputs(
             "mt",
             [
                 triggers.MTGenerateCrossTriggerFlags.output_group,
                 triggers.GenerateSingleTrailingTauTriggerFlags.output_group,
-                ] + [p for p in scalefactors.TauID_SF_v9.get_outputs("mt")
                 ] + [p for p in pairquantities.MTDiTauPairQuantities_v9.get_outputs("mt")
             ],
         )
@@ -2203,7 +1610,6 @@ def build_config(
             [
                 triggers.ETGenerateCrossTriggerFlags.output_group,
                 triggers.GenerateSingleTrailingTauTriggerFlags.output_group,
-                ] + [p for p in scalefactors.TauID_SF_v9.get_outputs("et")
                 ] + [p for p in pairquantities.ETDiTauPairQuantities_v9.get_outputs("et")
             ],
         )
@@ -2218,49 +1624,10 @@ def build_config(
             [
                 triggers.GenerateSingleTrailingTauTriggerFlags.output_group,
                 triggers.GenerateSingleLeadingTauTriggerFlags.output_group,
-                ] + [p for p in scalefactors.TauID_SF_v9.get_outputs("tt")
                 ] + [p for p in pairquantities.TTDiTauPairQuantities_v9.get_outputs("tt")
             ],
         )
     else:
-        if int(era[:4]) < 2024:
-            configuration.add_outputs(
-                "global",
-                [
-                    p for p in met.MetBasics_v12.get_outputs("global")
-                ],
-            )
-            configuration.add_outputs(
-                "mt",
-                [p for p in scalefactors.TauID_SF_v12.get_outputs("mt")],
-            )
-            configuration.add_outputs(
-                "et",
-                [p for p in scalefactors.TauID_SF_v12.get_outputs("et")],
-            )
-            configuration.add_outputs(
-                "tt",
-                [p for p in scalefactors.TauID_SF_v12.get_outputs("tt")],
-            )
-        else:
-            configuration.add_outputs(
-                "global",
-                [
-                    p for p in met.MetBasics.get_outputs("global")
-                ],
-            )
-            configuration.add_outputs(
-                "mt",
-                [p for p in scalefactors.TauID_SF.get_outputs("mt")],
-            )
-            configuration.add_outputs(
-                "et",
-                [p for p in scalefactors.TauID_SF.get_outputs("et")],
-            )
-            configuration.add_outputs(
-                "tt",
-                [p for p in scalefactors.TauID_SF.get_outputs("tt")],
-            )
         configuration.add_outputs(
             "mt",
             [
@@ -2293,7 +1660,6 @@ def build_config(
                 ] + [p for p in scalefactors.MuonIDIso_SF.get_outputs("em")
             ],
         )
-    
 
     if "data" not in sample and "embedding" not in sample:
         configuration.add_outputs(
@@ -2318,12 +1684,8 @@ def build_config(
         )
 
     if measure_btag_efficiency:
-        configuration.add_config_parameters(
-            "global",
-            {
-                "min_jet_pt_loose": 20,
-            },
-        )
+        configuration.add_config_parameters("global", {"min_jet_pt_loose": 20,},)
+
         if sample not in ["data", "embedding", "embedding_mc"]:
             configuration.add_producers(
                 scopes,
@@ -2343,429 +1705,7 @@ def build_config(
                     q.jet_btag_value_vec,
                 ],
             )
-
-    add_shift = get_adjusted_add_shift_SystematicShift(configuration)
-
-    #########################
-    # LHE Scale Weight variations
-    #########################
-    if "ggh" in sample or "qqh" in sample:
-        with defaults(scopes="global"):
-            with defaults(shift_map={"Up": 2.0, "Down": 0.5}):
-                add_shift(name="QCDscale_ren_ggH_qqH", shift_key="muR", producers=[event.LHE_Scale_weight])
-                add_shift(name="QCDscale_fac_ggH_qqH", shift_key="muF", producers=[event.LHE_Scale_weight])
-                add_shift(name="ps_fsr_ggH_qqH", shift_key="fsr", producers=[event.PS_weight])
-                add_shift(name="ps_isr_ggH_qqH", shift_key="isr", producers=[event.PS_weight])
-            with defaults(shift_map={"Up": "up", "Down": "down"}):
-                add_shift(name="pdf_Higgs_ggH_qqH", shift_key="pdf_variation", producers=[event.LHE_PDF_weight])
-                add_shift(name="pdf_alphas_ggH_qqH", shift_key="pdf_alphaS_variation", producers=[event.LHE_alphaS_weight])
-    
-    #########################
-    # Muon scale and resolution correction shifts
-    #########################
-    with defaults(
-        scopes="global",
-        shift_key="muon_sr_shift",
-        producers=[muons.MuonPtCorrection],
-        exclude_samples=["data", "embedding", "embedding_mc"],
-    ):
-        # Per systematics_master.yml (CMS_scale_m_<era>), muon energy scale is
-        # decorrelated across eras; era_tag is embedded in the name so each
-        # era's build_config invocation yields a distinct nuisance.
-        add_shift(name=f"CMS_scale_m_{era_tag}", shift_map={"Up": "ScaleUp", "Down": "ScaleDown"})
-        add_shift(name=f"CMS_res_m_{era_tag}", shift_map={"Up": "ResoUp", "Down": "ResoDown"})
-
-    #########################
-    # Electron energy correction shifts
-    #########################
-    if int(era[:4]) < 2022:
-        with defaults(
-            scopes="global",
-            shift_key="ele_es_variation",
-            producers=[electrons.ElectronPtCorrectionMC_v9],
-            exclude_samples=["data", "embedding", "embedding_mc"],
-        ):
-            # Per systematics_master.yml (CMS_res_e_<era>/CMS_scale_e_<era>),
-            # electron energy resolution/scale is decorrelated across eras.
-            add_shift(name=f"CMS_res_e_{era_tag}", shift_map={"Up": "resolutionUp", "Down": "resolutionDown"})
-            add_shift(name=f"CMS_scale_e_{era_tag}", shift_map={"Up": "scaleUp", "Down": "scaleDown"})
-    else:
-        with defaults(
-            scopes="global",
-            shift_key="ele_es_variation",
-            producers=[electrons.ElectronPtCorrectionMC],
-            exclude_samples=["data", "embedding", "embedding_mc"],
-        ):
-            # Per systematics_master.yml (CMS_res_e_<era>/CMS_scale_e_<era>),
-            # electron energy resolution/scale is decorrelated across eras.
-            add_shift(name=f"CMS_res_e_{era_tag}", shift_map={"Up": "resolutionUp", "Down": "resolutionDown"})
-            add_shift(name=f"CMS_scale_e_{era_tag}", shift_map={"Up": "scaleUp", "Down": "scaleDown"})
-
-    #########################
-    # Electron ID shifts
-    #########################
-    # Per systematics_master.yml (CMS_eff_e_id_<era>), the electron ID SF
-    # uncertainty is decorrelated across eras.
-    add_shift(
-        name=f"CMS_eff_e_id_{era_tag}",
-        scopes=("et", "ee", "em"),
-        shift_key="ele_sf_variation",
-        shift_map={"Up":"sfup", "Down":"sfdown"},
-        producers=[scalefactors.EleID_SF],
-    )
-
-    #########################
-    # MET Shifts
-    #########################
-    configuration.add_shift(
-        SystematicShiftByQuantity(
-            name="CMS_scale_met_unclustered_energyUp",
-            quantity_change={
-                nanoAODv15.PuppiMET_pt: "PuppiMET_ptUnclusteredUp",
-                nanoAODv15.PuppiMET_phi: "PuppiMET_phiUnclusteredUp",
-            },
-            scopes=["global"],
-        ),
-        exclude_samples=["data", "embedding", "embedding_mc"],
-    )
-    configuration.add_shift(
-        SystematicShiftByQuantity(
-            name="CMS_scale_met_unclustered_energyDown",
-            quantity_change={
-                nanoAODv15.PuppiMET_pt: "PuppiMET_ptUnclusteredDown",
-                nanoAODv15.PuppiMET_phi: "PuppiMET_phiUnclusteredDown",
-            },
-            scopes=["global"],
-        ),
-        exclude_samples=["data", "embedding", "embedding_mc"],
-    )
-
-    #########################
-    # MET Recoil Shifts
-    #########################
-    if int(era[:4]) < 2022:
-        with defaults(
-            scopes=("et", "mt", "tt", "em", "ee", "mm"),
-            producers=[met.ApplyRecoilCorrections_Run2],
-            exclude_samples=["data", "embedding", "embedding_mc"],
-            shift_key=[
-                "apply_recoil_resolution_systematic",  # set either to True or False
-                "apply_recoil_response_systematic",  # set either to True or False
-                "recoil_systematic_shift_up",  # set either to True or False upon variation
-                "recoil_systematic_shift_down",  # set either to True or False upon variation
-            ]
-        ):
-            add_shift(
-                name="CMS_scale_met_recoil_response",
-                shift_map={
-                    "Up": [False, True, True, False],
-                    "Down": [False, True, False, True],
-                }
-            )
-            add_shift(
-                name="CMS_res_met_recoil_resolution",
-                shift_map={
-                    "Up": [True, False, True, False],
-                    "Down": [True, False, False, True],
-                },
-            )
-    else:
-        # NOTE: per the HLepRare recoil documentation, the Response/Resolution
-        # uncertainty must be evaluated on the MET *after* the nominal
-        # (QuantileMapHist) recoil correction has already been applied (H_para/
-        # H_perp are computed from the QuantileMapHist-corrected MET, see the
-        # "Uncertainty description" / example snippet on the recoil corrections
-        # page). We therefore only toggle the dedicated ApplyRecoilUncertainty
-        # producer (chained after met.ApplyRecoilCorrections, which always keeps
-        # using the nominal "QuantileMapHist" method) instead of switching the
-        # method of ApplyRecoilCorrections itself, which would incorrectly
-        # evaluate the uncertainty on the pre-recoil-correction MET.
-        with defaults(
-            scopes=("et", "mt", "tt", "em", "ee", "mm"),
-            producers=[met.ApplyRecoilUncertainty],
-            exclude_samples=["data", "embedding", "embedding_mc"],
-            shift_key=["applyRecoilUncertainty", "recoil_uncertainty_variation"]
-        ):
-            add_shift(
-                name="CMS_scale_met_recoil_response",
-                shift_map={"Up": [True, "RespUp"], "Down": [True, "RespDown"]}
-            )
-            add_shift(
-                name="CMS_res_met_recoil_resolution",
-                shift_map={"Up": [True, "ResolUp"], "Down": [True, "ResolDown"]}
-            )
-
-    #########################
-    # Pileup Shifts
-    #########################
-    # if era == "2025":
-    #     add_shift(
-    #         name=f"PileUp",
-    #         shift_key="PU_reweighting_variation",
-    #         shift_map={"Up": "data/root_pileup/Data_PileUp_2025_72p3832.root", "Down": "data/root_pileup/Data_PileUp_2025_66p0168.root"},
-    #         scopes="global",
-    #         producers=[event.PUweights],
-    #         exclude_samples=["data", "embedding", "embedding_mc"],
-    #     )
-    # else:
-    # Per systematics_master.yml (CMS_pileup_<era>), pileup reweighting is
-    # decorrelated across eras.
-    add_shift(
-        name=f"CMS_pileup_{era_tag}",
-        shift_key="PU_reweighting_file_data",
-        shift_map={"Up": "up", "Down": "down"},
-        scopes="global",
-        producers=[event.PUweights],
-        exclude_samples=["data", "embedding", "embedding_mc"],
-    )
-
-    #########################
-    # Prefiring Shifts
-    #########################
-    if int(era[:4]) < 2018:
-        configuration.add_shift(
-            SystematicShiftByQuantity(
-                name="CMS_l1_ecal_prefiringDown",
-                quantity_change={
-                    nanoAODv9.L1PreFiringWeight_Nom: "L1PreFiringWeight_Dn",
-                },
-                scopes=["global"],
-            )
-        )
-        configuration.add_shift(
-            SystematicShiftByQuantity(
-                name="CMS_l1_ecal_prefiringUp",
-                quantity_change={
-                    nanoAODv9.L1PreFiringWeight_Nom: "L1PreFiringWeight_Up",
-                },
-                scopes=["global"],
-            )
-        )
-
-    #########################
-    # Z pt DY uncertainties
-    #########################
-    # Per the CMS H->tautau DY kinematic reweighting recommendation
-    # (https://cms-higgs-leprare.docs.cern.ch/htt-common/DY_reweight/), the
-    # DY_pTll_reweighting correction provides N fit-parameter uncertainties per
-    # DY sample order (8 for LO/madgraph, 10 for NLO/amcatnlo, 9 for NNLO/powheg,
-    # also given by the DY_pTll_reweighting_N_uncertainty correction). The doc
-    # states: "all uncertainties are required to be included and are fully
-    # uncorrelated from each other." A single up/down envelope therefore does not
-    # reflect the recommendation; instead one fully-uncorrelated shift is added
-    # per fit-parameter uncertainty ("upN"/"downN"), split by DY sample order so
-    # that only the number of uncertainties valid for that order is used.
-    # Each per-fit-parameter uncertainty's name embeds era_tag so that, in
-    # addition to being uncorrelated from each other (see above), they are
-    # also uncorrelated across eras: without the era tag, "zPtReweightWeight_
-    # unc1" from every era's build_config invocation would collide into a
-    # single, unintentionally era-correlated nuisance name downstream.
-    DY_PTLL_N_UNC = {"LO": 8, "NLO": 10, "NNLO": 9}
-    if int(era[:4]) >= 2022:
-        # NNLO (powheg) samples
-        for n in range(1, DY_PTLL_N_UNC["NNLO"] + 1):
-            add_shift(
-                scopes=("et", "mt", "tt", "em", "ee", "mm"),
-                producers=[event.ZPtReweighting],
-                shift_key="zpt_variation",
-                name=f"CMS_htt_zptReweight_unc{n}_{era_tag}",
-                shift_map={"Up": f"up{n}", "Down": f"down{n}"},
-                samples=["dyjets_powheg"],
-            )
-        # NLO (amcatnlo) samples, incl. combined electroweak_boson sample
-        for n in range(1, DY_PTLL_N_UNC["NLO"] + 1):
-            add_shift(
-                scopes=("et", "mt", "tt", "em", "ee", "mm"),
-                producers=[event.ZPtReweighting],
-                shift_key="zpt_variation",
-                name=f"CMS_htt_zptReweight_unc{n}_{era_tag}",
-                shift_map={"Up": f"up{n}", "Down": f"down{n}"},
-                samples=["dyjets_amcatnlo", "dyjets_amcatnlo_ll", "dyjets_amcatnlo_tt", "electroweak_boson"],
-            )
-    
-    #########################
-    # Tau energy scale shifts  #
-    #########################
-    with defaults(shift_map={"Down": "down", "Up": "up"}):
-        if ("dyjets" in sample or "electroweak_boson" in sample) and int(era[:4]) < 2022:
-            add_shift(
-                name="CMS_scale_t_genMuon",
-                shift_key="tau_mufake_es",
-                scopes="mt",
-                producers=[taus.TauPtCorrection_muFake],
-            )
-            with defaults(
-                scopes="et",
-                producers=[taus.TauPtCorrection_eleFake],
-            ):
-                add_shift(name="CMS_scale_t_DM0_genElectron_barrel", shift_key="tau_elefake_es_DM0_barrel")
-                add_shift(name="CMS_scale_t_DM0_genElectron_endcap", shift_key="tau_elefake_es_DM0_endcap")
-                add_shift(name="CMS_scale_t_DM1_genElectron_barrel", shift_key="tau_elefake_es_DM1_barrel")
-                add_shift(name="CMS_scale_t_DM1_genElectron_endcap", shift_key="tau_elefake_es_DM1_endcap")
-        elif int(era[:4]) < 2022:
-            with defaults(scopes=("et", "mt", "tt")):
-                # dm and pt scheme
-                with defaults(producers=[taus.TauEnergyCorrection_ES_dm_pt_binned]):
-                    for dm, dm_num in [
-                        ("1prong0pizero", "0"),
-                        ("1prong1pizero", "1"),
-                        ("3prong0pizero", "10"),
-                        ("3prong1pizero", "11"),
-                    ]:
-                        for pt in ["20to40", "40toInf"]:
-                            add_shift(name=f"CMS_scale_t_DM{dm_num}_genTau_pT{pt}", shift_key=f"tau_ES_shift_{dm}{pt}")
-
-        elif int(era[:4]) >= 2022 and int(era[:4]) < 2024:
-            with defaults(scopes=("et", "mt", "tt")): #is there a reason not to apply this everywhere?
-                with defaults(producers=[taus.TauEnergyCorrection_v12]): # propagate to mass too
-                    for dm in ["0", "1", "10", "11"]:
-                        # genuine tau
-                        add_shift(name=f"CMS_scale_t_DM{dm}_genTau_{era_tag}", shift_key=f"tau_es_DM{dm}")
-                        # ele fake
-                        add_shift(name=f"CMS_scale_t_DM{dm}_genElectron_barrel_{era_tag}", shift_key=f"tau_elefake_es_DM{dm}_barrel")
-                        add_shift(name=f"CMS_scale_t_DM{dm}_genElectron_endcap_{era_tag}", shift_key=f"tau_elefake_es_DM{dm}_endcap")
-                        # muon fake
-                        add_shift(name=f"CMS_scale_t_DM{dm}_genMuon_{era_tag}", shift_key=f"tau_mufake_es_DM{dm}")
-        elif int(era[:4]) >= 2024:
-            with defaults(scopes=("et", "mt", "tt")): #is there a reason not to apply this everywhere?
-                with defaults(producers=[taus.TauEnergyCorrection]): # propagate to mass too
-                    for dm in ["0", "1", "10", "11"]:
-                        # genuine tau
-                        for pt in ["20to40", "40to60", "60toInf"]:
-                            add_shift(name=f"CMS_scale_t_DM{dm}_genTau_pT{pt}_{era_tag}", shift_key=f"tau_es_DM{dm}_pt{pt}")
-                        # ele fake
-                        add_shift(name=f"CMS_scale_t_DM{dm}_genElectron_barrel_{era_tag}", shift_key=f"tau_elefake_es_DM{dm}_barrel")
-                        add_shift(name=f"CMS_scale_t_DM{dm}_genElectron_endcap_{era_tag}", shift_key=f"tau_elefake_es_DM{dm}_endcap")
-                        # muon fake
-                        add_shift(name=f"CMS_scale_t_DM{dm}_genMuon_{era_tag}", shift_key=f"tau_mufake_es_DM{dm}")
-
-    #########################
-    # TauID scale factor shifts
-    #########################
-    with defaults(
-        exclude_samples=["data", "embedding", "embedding_mc"],
-        shift_map={"Up": "up", "Down": "down"}
-        ):
-
-        if int(era[:4]) < 2022:
-            with defaults(scopes=("et", "mt")):
-                #dm and pt scheme
-                with defaults(producers=[scalefactors.Tau_2_VsJetTauID_lt_SF_dm_pt_binned]):
-                    for dm, dm_num in [
-                        ("1prong0pizero", "0"),
-                        ("1prong1pizero", "1"),
-                        ("3prong0pizero", "10"),
-                        ("3prong1pizero", "11"),
-                    ]:
-                        for pt in ["20to40", "40toInf"]:
-                            add_shift(name=f"CMS_eff_t_DeepTau2017v2p1_VSjet_DM{dm_num}_pT{pt}", shift_key=f"tau_id_vsjet_{dm}{pt}")
-                with defaults(producers=[scalefactors.Tau_2_VsEleTauID_SF_Run2]):
-                    add_shift(name="CMS_fake_t_DeepTau2017v2p1_VSe_barrel", shift_key="tau_id_vsele_barrel")
-                    add_shift(name="CMS_fake_t_DeepTau2017v2p1_VSe_endcap", shift_key="tau_id_vsele_endcap")
-                with defaults(producers=[scalefactors.Tau_2_VsMuTauID_SF]):
-                    for wheel in range(1, 6):
-                        add_shift(name=f"CMS_fake_t_DeepTau2017v2p1_VSmu_wheel{wheel}", shift_key=f"tau_id_vsmu_wheel{wheel}")
-            with defaults(scopes="tt"):
-                with defaults(producers=[scalefactors.Tau_1_VsJetTauID_SF_v12, scalefactors.Tau_2_VsJetTauID_tt_SF]):
-                    add_shift(name="CMS_eff_t_DeepTau2017v2p1_VSjet_DM0", shift_key="tau_id_vsjet_DM0")
-                    add_shift(name="CMS_eff_t_DeepTau2017v2p1_VSjet_DM1", shift_key="tau_id_vsjet_DM1")
-                    add_shift(name="CMS_eff_t_DeepTau2017v2p1_VSjet_DM10", shift_key="tau_id_vsjet_DM10")
-                    add_shift(name="CMS_eff_t_DeepTau2017v2p1_VSjet_DM11", shift_key="tau_id_vsjet_DM11")
-                with defaults(producers=[scalefactors.Tau_1_VsEleTauID_SF_Run2, scalefactors.Tau_2_VsEleTauID_SF_Run2]):
-                    add_shift(name="CMS_fake_t_DeepTau2017v2p1_VSe_barrel", shift_key="tau_id_vsele_barrel")
-                    add_shift(name="CMS_fake_t_DeepTau2017v2p1_VSe_endcap", shift_key="tau_id_vsele_endcap")
-                with defaults(producers=[scalefactors.Tau_1_VsMuTauID_SF, scalefactors.Tau_2_VsMuTauID_SF]):
-                    for wheel in range(1, 6):
-                        add_shift(name=f"CMS_fake_t_DeepTau2017v2p1_VSmu_wheel{wheel}", shift_key=f"tau_id_vsmu_wheel{wheel}")
-            
-        else:
-            for dm in ["0", "1", "10", "11"]:
-                # vs Ele
-                with defaults(name=f"CMS_fake_t_DeepTau2018v2p5_VSe_DM{dm}_barrel_{era_tag}", shift_key=f"tau_id_vsele_DM{dm}_barrel"):
-                    add_shift(scopes=("et", "mt", "tt"),producers=[scalefactors.Tau_2_VsEleTauID_SF])
-                    add_shift(scopes=("tt"),producers=[scalefactors.Tau_1_VsEleTauID_SF])
-                with defaults(name=f"CMS_fake_t_DeepTau2018v2p5_VSe_DM{dm}_endcap_{era_tag}", shift_key=f"tau_id_vsele_DM{dm}_endcap"):
-                    add_shift(scopes=("et", "mt", "tt"),producers=[scalefactors.Tau_2_VsEleTauID_SF])
-                    add_shift(scopes=("tt"),producers=[scalefactors.Tau_1_VsEleTauID_SF])
-                # vs Jet
-                if int(era[:4]) < 2024:
-                    # NOTE: for eras before 2024 the "DeepTau2018v2p5VSjet"
-                    # correctionlib json (flag="dm") does *not* provide plain
-                    # "nom"/"up"/"down" categories for the "syst" input. It
-                    # instead only defines the individual, named uncertainty
-                    # sources (stat1_dm<DM>, stat2_dm<DM>, syst_alleras,
-                    # syst_<era>, syst_TES_<era>_dm<DM>) with a "default"
-                    # fallback formula that returns the nominal SF for *any*
-                    # unmatched string, including "up"/"down". Passing plain
-                    # "up"/"down" (as previously done here) therefore silently
-                    # evaluated to the nominal SF for both directions, i.e. the
-                    # systematic had no effect at all. We now pass the exact
-                    # json keys for each individual uncertainty source, as
-                    # documented on the TauPOG Wiki ("Run 2" recipe, reused
-                    # for 2022/2023): https://tau-wiki.docs.cern.ch (higher
-                    # precedence than the outdated CMS TWiki TauIDRecommendationForRun3 page).
-                    era_key = era[:4] + "_" + era[4:]  # e.g. 2022preEE -> 2022_preEE
-                    with defaults(shift_key=f"tau_id_vsjet_DM{dm}"):
-                        with defaults(
-                            # stat uncertainties are decorrelated per DM *and*
-                            # per era (2 x 4 x 4 = 16 nuisances in total, per
-                            # the TauPOG Wiki), hence era is part of the name.
-                            name=f"vsJetStat1{era}DM{dm}",
-                            shift_map={"Up": f"stat1_dm{dm}_up", "Down": f"stat1_dm{dm}_down"},
-                        ):
-                            add_shift(scopes=("et", "mt", "tt"), producers=[scalefactors.Tau_2_VsJetTauID_SF_v12])
-                            add_shift(scopes=("tt"), producers=[scalefactors.Tau_1_VsJetTauID_SF_v12])
-                        with defaults(
-                            name=f"vsJetStat2{era}DM{dm}",
-                            shift_map={"Up": f"stat2_dm{dm}_up", "Down": f"stat2_dm{dm}_down"},
-                        ):
-                            add_shift(scopes=("et", "mt", "tt"), producers=[scalefactors.Tau_2_VsJetTauID_SF_v12])
-                            add_shift(scopes=("tt"), producers=[scalefactors.Tau_1_VsJetTauID_SF_v12])
-                        with defaults(
-                            name=f"vsJetSystTES{era}DM{dm}",
-                            shift_map={
-                                "Up": f"syst_TES_{era_key}_dm{dm}_up",
-                                "Down": f"syst_TES_{era_key}_dm{dm}_down",
-                            },
-                        ):
-                            add_shift(scopes=("et", "mt", "tt"), producers=[scalefactors.Tau_2_VsJetTauID_SF_v12])
-                            add_shift(scopes=("tt"), producers=[scalefactors.Tau_1_VsJetTauID_SF_v12])
-                else:
-                    for pt in ["20to40", "40to60", "60toInf"]:
-                        with defaults(name=f"CMS_eff_t_DeepTau2018v2p5_VSjet_DM{dm}_pT{pt}", shift_key=f"tau_id_vsjet_DM{dm}_pt{pt}"):
-                            add_shift(scopes=("et", "mt", "tt"),producers=[scalefactors.Tau_2_VsJetTauID_SF])
-                            add_shift(scopes=("tt"),producers=[scalefactors.Tau_1_VsJetTauID_SF])
-            if int(era[:4]) < 2024:
-                # syst_alleras: fully correlated across DMs and (by construction
-                # of reusing the same shift name in every era's config) across
-                # eras as well.
-                era_key = era[:4] + "_" + era[4:]
-                dm_keys = ["tau_id_vsjet_DM0", "tau_id_vsjet_DM1", "tau_id_vsjet_DM10", "tau_id_vsjet_DM11"]
-                with defaults(
-                    name="vsJetSystAllEras",
-                    shift_key=dm_keys,
-                    shift_map={"Up": ["syst_alleras_up"] * 4, "Down": ["syst_alleras_down"] * 4},
-                ):
-                    add_shift(scopes=("et", "mt", "tt"), producers=[scalefactors.Tau_2_VsJetTauID_SF_v12])
-                    add_shift(scopes=("tt"), producers=[scalefactors.Tau_1_VsJetTauID_SF_v12])
-                # syst_<era>: correlated across DMs, uncorrelated across eras
-                # (implicit since era_key/name differ per era's config).
-                with defaults(
-                    name=f"vsJetSyst{era}",
-                    shift_key=dm_keys,
-                    shift_map={
-                        "Up": [f"syst_{era_key}_up"] * 4,
-                        "Down": [f"syst_{era_key}_down"] * 4,
-                    },
-                ):
-                    add_shift(scopes=("et", "mt", "tt"), producers=[scalefactors.Tau_2_VsJetTauID_SF_v12])
-                    add_shift(scopes=("tt"), producers=[scalefactors.Tau_1_VsJetTauID_SF_v12])
-            # vs Muon
-            for wheel in range(1, 6):
-                with defaults(name=f"CMS_fake_t_DeepTau2018v2p5_VSmu_wheel{wheel}_{era_tag}", shift_key=f"tau_id_vsmu_wheel{wheel}"):
-                    add_shift(scopes=("et", "mt", "tt"),producers=[scalefactors.Tau_2_VsMuTauID_SF])
-                    add_shift(scopes=("tt"),producers=[scalefactors.Tau_1_VsMuTauID_SF])
-
+ 
     #########################
     # Add additional producers and SFs related to embedded samples
     #########################
@@ -2778,9 +1718,9 @@ def build_config(
     configuration = add_diTauTriggerSetup(configuration)
 
     #########################
-    # Jet energy resolution and jet energy scale and btag uncertainties
+    # Systematics shifts
     #########################
-    configuration = add_jetVariations(configuration, era)
+    configuration = add_Variations(configuration, sample, era)
 
     #########################
     # Finalize and validate the configuration
