+++ tau_triggersetup.py
@@ -1,8 +1,67 @@
 from code_generation.configuration import Configuration
 from code_generation.modifiers import EraModifier, SampleModifier
-from code_generation.systematics import SystematicShift, SystematicShiftByQuantity
-from .scripts.CROWNWrapper import defaults, get_adjusted_add_shift_SystematicShift
-from .producers import scalefactors as scalefactors
+
+MUTAU_CROSS_TRIGGER_FLAG = EraModifier(
+    {
+        "2016preVFP": "trg_cross_mu20tau27_hps",
+        "2016postVFP": "trg_cross_mu20tau27_hps",
+        "2017": "trg_cross_mu20tau27_hps",
+        "2018": "trg_cross_mu20tau27_hps",
+        "2022preEE": "trg_cross_mu20tau27_hps",
+        "2022postEE": "trg_cross_mu20tau27_hps",
+        "2023preBPix": "trg_cross_mu20tau27_hps",
+        "2023postBPix": "trg_cross_mu20tau27_hps",
+        "2024": "trg_cross_mu20tau27_hps",
+        "2025": "trg_cross_mu20tau27_pnet",
+        "2026": "trg_cross_mu20tau27_pnet",
+    }
+)
+
+ELETAU_CROSS_TRIGGER_FLAG = EraModifier(
+    {
+        "2016preVFP": "trg_cross_ele24tau30_hps",
+        "2016postVFP": "trg_cross_ele24tau30_hps",
+        "2017": "trg_cross_ele24tau30_hps",
+        "2018": "trg_cross_ele24tau30_hps",
+        "2022preEE": "trg_cross_ele24tau30_hps",
+        "2022postEE": "trg_cross_ele24tau30_hps",
+        "2023preBPix": "trg_cross_ele24tau30_hps",
+        "2023postBPix": "trg_cross_ele24tau30_hps",
+        "2024": "trg_cross_ele24tau30_hps",
+        "2025": "trg_cross_ele24tau30_pnet",
+        "2026": "trg_cross_ele24tau30_pnet",
+    }
+)
+
+RUN2_ERAS = ["2016preVFP", "2016postVFP", "2017", "2018"]
+DOUBLETAU_HPS_ERAS = ["2022preEE", "2022postEE", "2023preBPix", "2023postBPix"]
+
+# hps->pnet transition in 2025, shared by both legs and by nominal/shifted ditau SF blocks
+DOUBLETAU_TRIGGER_LEG1_FLAGNAME = EraModifier(
+    {
+        **{era: '""' for era in RUN2_ERAS},
+        **{era: "trg_wgt_doubletau35_leg1" for era in DOUBLETAU_HPS_ERAS},
+    },
+    default="trg_wgt_doubletau30_leg1",  # 2024, 2025, 2026
+)
+DOUBLETAU_TRIGGER_LEG2_FLAGNAME = EraModifier(
+    {
+        **{era: '""' for era in RUN2_ERAS},
+        **{era: "trg_wgt_doubletau35_leg2" for era in DOUBLETAU_HPS_ERAS},
+    },
+    default="trg_wgt_doubletau30_leg2",  # 2024, 2025, 2026
+)
+DOUBLETAU_TRIGGER_FLAG = EraModifier(
+    {
+        **{era: '""' for era in RUN2_ERAS},
+        **{era: "trg_double_tau35_mediumiso_hps" for era in DOUBLETAU_HPS_ERAS},
+    },
+    default="trg_double_tau30_mediumiso_pnet",  # 2024, 2025, 2026
+)
+DOUBLETAU_TRIGGER_SF_NAME = EraModifier(
+    {era: '""' for era in RUN2_ERAS},
+    default="ditau",  # all Run 3 eras
+)
 
 
 def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
@@ -17,81 +76,23 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
         "max_deltaR_triggermatch": 0.4,
     }
     
+    # shared by all Run 3 eras (2022preEE through 2026)
+    singlemuon_trigger_run3_default = [
+        {
+            "flagname": "trg_single_mu24",
+            "hlt_path": "HLT_IsoMu24",
+            "ptcut": 26,
+            "etacut": 2.4,
+            "filterbit": 1,
+            **singlemuon_trigger_defaults,
+        },
+    ]
+
     configuration.add_config_parameters(
         ["mt", "mm", "em"],
         {
             "singlemuon_trigger": EraModifier(
                 {
-                    "2026": [
-                        {
-                            "flagname": "trg_single_mu24",
-                            "hlt_path": "HLT_IsoMu24",
-                            "ptcut": 26,
-                            "etacut": 2.4,
-                            "filterbit": 1,
-                            **singlemuon_trigger_defaults,
-                        },
-                    ],
-                    "2025": [
-                        {
-                            "flagname": "trg_single_mu24",
-                            "hlt_path": "HLT_IsoMu24",
-                            "ptcut": 26,
-                            "etacut": 2.4,
-                            "filterbit": 1,
-                            **singlemuon_trigger_defaults,
-                        },
-                    ],
-                    "2024": [
-                        {
-                            "flagname": "trg_single_mu24",
-                            "hlt_path": "HLT_IsoMu24",
-                            "ptcut": 26,
-                            "etacut": 2.4,
-                            "filterbit": 1,
-                            **singlemuon_trigger_defaults,
-                        },
-                    ],
-                    "2023postBPix": [
-                        {
-                            "flagname": "trg_single_mu24",
-                            "hlt_path": "HLT_IsoMu24",
-                            "ptcut": 26,
-                            "etacut": 2.4,
-                            "filterbit": 1,
-                            **singlemuon_trigger_defaults,
-                        },
-                    ],
-                    "2023preBPix": [
-                        {
-                            "flagname": "trg_single_mu24",
-                            "hlt_path": "HLT_IsoMu24",
-                            "ptcut": 26,
-                            "etacut": 2.4,
-                            "filterbit": 1,
-                            **singlemuon_trigger_defaults,
-                        },
-                    ],
-                    "2022postEE": [
-                        {
-                            "flagname": "trg_single_mu24",
-                            "hlt_path": "HLT_IsoMu24",
-                            "ptcut": 26,
-                            "etacut": 2.4,
-                            "filterbit": 1,
-                            **singlemuon_trigger_defaults,
-                        },
-                    ],
-                    "2022preEE": [
-                        {
-                            "flagname": "trg_single_mu24",
-                            "hlt_path": "HLT_IsoMu24",
-                            "ptcut": 26,
-                            "etacut": 2.4,
-                            "filterbit": 1,
-                            **singlemuon_trigger_defaults,
-                        },
-                    ],
                     "2018": [
                         {
                             "flagname": "trg_single_mu24",
@@ -212,7 +213,8 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
                             **singlemuon_trigger_defaults,
                         },
                     ],
-                }
+                },
+                default=singlemuon_trigger_run3_default,
             ),
         },
     )
@@ -230,28 +232,20 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
         {
             "mutau_cross_trigger": EraModifier(
                 {
-                    "2026": [
-                        {
-                            "flagname": "trg_cross_mu20tau27_pnet",
-                            "hlt_path": "HLT_IsoMu20_eta2p1_PNetTauhPFJet27_Medium_eta2p3_CrossL1",
-                            "p1_ptcut": 20,
-                            "p2_ptcut": 32,
-                            "p1_filterbit": 1,
-                            "p2_filterbit": "1, 4, 13",
-                            **mutau_cross_trigger_defaults,
-                        },
-                    ],
-                    "2025": [
-                        {
-                            "flagname": "trg_cross_mu20tau27_pnet",
-                            "hlt_path": "HLT_IsoMu20_eta2p1_PNetTauhPFJet27_Medium_eta2p3_CrossL1",
-                            "p1_ptcut": 20,
-                            "p2_ptcut": 32,
-                            "p1_filterbit": 1,
-                            "p2_filterbit": "1, 4, 13",
-                            **mutau_cross_trigger_defaults,
-                        },
-                    ],
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_cross_mu20tau27_pnet",
+                                "hlt_path": "HLT_IsoMu20_eta2p1_PNetTauhPFJet27_Medium_eta2p3_CrossL1",
+                                "p1_ptcut": 20,
+                                "p2_ptcut": 32,
+                                "p1_filterbit": 1,
+                                "p2_filterbit": "1, 4, 13",
+                                **mutau_cross_trigger_defaults,
+                            },
+                        ]
+                        for era in ["2025", "2026"]
+                    },
                     "2024": [
                         {
                             "flagname": "trg_cross_mu20tau27_hps",
@@ -259,51 +253,7 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
                             "p1_ptcut": 20,
                             "p2_ptcut": 32,
                             "p1_filterbit": 1,
-                            "p2_filterbit": "3, 13",  
-                            **mutau_cross_trigger_defaults,
-                        },
-                    ],
-                    "2023postBPix": [
-                        {
-                            "flagname": "trg_cross_mu20tau27_hps",
-                            "hlt_path": "HLT_IsoMu20_eta2p1_LooseDeepTauPFTauHPS27_eta2p1_CrossL1",
-                            "p1_ptcut": 20,
-                            "p2_ptcut": 32,
-                            "p1_filterbit": 1,
-                            "p2_filterbit": "3, 9",  
-                            **mutau_cross_trigger_defaults,
-                        },
-                    ],
-                    "2023preBPix": [
-                        {
-                            "flagname": "trg_cross_mu20tau27_hps",
-                            "hlt_path": "HLT_IsoMu20_eta2p1_LooseDeepTauPFTauHPS27_eta2p1_CrossL1",
-                            "p1_ptcut": 20,
-                            "p2_ptcut": 32,
-                            "p1_filterbit": 1,
-                            "p2_filterbit": "3, 9",  
-                            **mutau_cross_trigger_defaults,
-                        },
-                    ],
-                    "2022postEE": [
-                        {
-                            "flagname": "trg_cross_mu20tau27_hps",
-                            "hlt_path": "HLT_IsoMu20_eta2p1_LooseDeepTauPFTauHPS27_eta2p1_CrossL1",
-                            "p1_ptcut": 20,
-                            "p2_ptcut": 32,
-                            "p1_filterbit": 1,
-                            "p2_filterbit": "3, 9",  
-                            **mutau_cross_trigger_defaults,
-                        },
-                    ],
-                    "2022preEE": [
-                        {
-                            "flagname": "trg_cross_mu20tau27_hps",
-                            "hlt_path": "HLT_IsoMu20_eta2p1_LooseDeepTauPFTauHPS27_eta2p1_CrossL1",
-                            "p1_ptcut": 20,
-                            "p2_ptcut": 32,
-                            "p1_filterbit": 1,
-                            "p2_filterbit": "3, 9",  
+                            "p2_filterbit": "3, 13",
                             **mutau_cross_trigger_defaults,
                         },
                     ],
@@ -314,7 +264,7 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
                             "p1_ptcut": 21,
                             "p2_ptcut": 32,
                             "p1_filterbit": 3,
-                            "p2_filterbit": 4, 
+                            "p2_filterbit": 4,
                             **mutau_cross_trigger_defaults,
                         },
                         # the non HPS version exists for data only, but add it anyway to have the flag in the ntuple
@@ -324,7 +274,7 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
                             "p1_ptcut": 21,
                             "p2_ptcut": 32,
                             "p1_filterbit": 3,
-                            "p2_filterbit": 4, 
+                            "p2_filterbit": 4,
                             **mutau_cross_trigger_defaults,
                         },
                     ],
@@ -335,33 +285,36 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
                             "p1_ptcut": 21,
                             "p2_ptcut": 32,
                             "p1_filterbit": 3,
-                            "p2_filterbit": 4, 
+                            "p2_filterbit": 4,
                             **mutau_cross_trigger_defaults,
                         }
                     ],
-                    "2016postVFP": [
-                        {
-                            "flagname": "trg_cross_mu19tau20",
-                            "hlt_path": "HLT_IsoMu19_eta2p1_LooseIsoPFTau20_SingleL1",
-                            "p1_ptcut": 20,
-                            "p2_ptcut": 22,
-                            "p1_filterbit": 3,
-                            "p2_filterbit": 4, 
-                            **mutau_cross_trigger_defaults,
-                        }
-                    ],
-                    "2016preVFP": [
-                        {
-                            "flagname": "trg_cross_mu19tau20",
-                            "hlt_path": "HLT_IsoMu19_eta2p1_LooseIsoPFTau20_SingleL1",
-                            "p1_ptcut": 20,
-                            "p2_ptcut": 22,
-                            "p1_filterbit": 3,
-                            "p2_filterbit": 4, 
-                            **mutau_cross_trigger_defaults,
-                        }
-                    ],
-                }
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_cross_mu19tau20",
+                                "hlt_path": "HLT_IsoMu19_eta2p1_LooseIsoPFTau20_SingleL1",
+                                "p1_ptcut": 20,
+                                "p2_ptcut": 22,
+                                "p1_filterbit": 3,
+                                "p2_filterbit": 4,
+                                **mutau_cross_trigger_defaults,
+                            }
+                        ]
+                        for era in ["2016postVFP", "2016preVFP"]
+                    },
+                },
+                default=[  # 2022preEE, 2022postEE, 2023preBPix, 2023postBPix
+                    {
+                        "flagname": "trg_cross_mu20tau27_hps",
+                        "hlt_path": "HLT_IsoMu20_eta2p1_LooseDeepTauPFTauHPS27_eta2p1_CrossL1",
+                        "p1_ptcut": 20,
+                        "p2_ptcut": 32,
+                        "p1_filterbit": 1,
+                        "p2_filterbit": "3, 9",
+                        **mutau_cross_trigger_defaults,
+                    },
+                ],
             ),
         },
     )
@@ -378,134 +331,54 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
         {
             "singleelectron_trigger": EraModifier(
                 {
-                    "2026": [
-                        {
-                            "flagname": "trg_single_ele30",
-                            "hlt_path": "HLT_Ele30_WPTight_Gsf",
-                            "ptcut": 32,
-                            "etacut": 2.5,
-                            **singleelectron_trigger_defaults,
-                        },
-                    ],
-                    "2025": [
-                        {
-                            "flagname": "trg_single_ele30",
-                            "hlt_path": "HLT_Ele30_WPTight_Gsf",
-                            "ptcut": 32,
-                            "etacut": 2.5,
-                            **singleelectron_trigger_defaults,
-                        },
-                    ],
-                    "2024": [
-                        {
-                            "flagname": "trg_single_ele30",
-                            "hlt_path": "HLT_Ele30_WPTight_Gsf",
-                            "ptcut": 32,
-                            "etacut": 2.5,
-                            **singleelectron_trigger_defaults,
-                        },
-                    ],
-                    "2023postBPix": [
-                        {
-                            "flagname": "trg_single_ele30",
-                            "hlt_path": "HLT_Ele30_WPTight_Gsf",
-                            "ptcut": 32,
-                            "etacut": 2.5,
-                            **singleelectron_trigger_defaults,
-                        },
-                    ],
-                    "2023preBPix": [
-                        {
-                            "flagname": "trg_single_ele30",
-                            "hlt_path": "HLT_Ele30_WPTight_Gsf",
-                            "ptcut": 32,
-                            "etacut": 2.5,
-                            **singleelectron_trigger_defaults,
-                        },
-                    ],
-                    "2022postEE": [
-                        {
-                            "flagname": "trg_single_ele30",
-                            "hlt_path": "HLT_Ele30_WPTight_Gsf",
-                            "ptcut": 32,
-                            "etacut": 2.5,
-                            **singleelectron_trigger_defaults,
-                        },
-                    ],
-                    "2022preEE": [
-                        {
-                            "flagname": "trg_single_ele30",
-                            "hlt_path": "HLT_Ele30_WPTight_Gsf",
-                            "ptcut": 32,
-                            "etacut": 2.5,
-                            **singleelectron_trigger_defaults,
-                        },
-                    ],
-                    "2018": [
-                        {
-                            "flagname": "trg_single_ele27",
-                            "hlt_path": "HLT_Ele27_WPTight_Gsf",
-                            "ptcut": 28,
-                            "etacut": 2.1,
-                            **singleelectron_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_single_ele32",
-                            "hlt_path": "HLT_Ele32_WPTight_Gsf",
-                            "ptcut": 33,
-                            "etacut": 2.1,
-                            **singleelectron_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_single_ele35",
-                            "hlt_path": "HLT_Ele35_WPTight_Gsf",
-                            "ptcut": 36,
-                            "etacut": 2.1,
-                            **singleelectron_trigger_defaults,
-                        },
-                    ],
-                    "2017": [
-                        {
-                            "flagname": "trg_single_ele27",
-                            "hlt_path": "HLT_Ele27_WPTight_Gsf",
-                            "ptcut": 28,
-                            "etacut": 2.1,
-                            **singleelectron_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_single_ele32",
-                            "hlt_path": "HLT_Ele32_WPTight_Gsf",
-                            "ptcut": 33,
-                            "etacut": 2.1,
-                            **singleelectron_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_single_ele35",
-                            "hlt_path": "HLT_Ele35_WPTight_Gsf",
-                            "ptcut": 36,
-                            "etacut": 2.1,
-                            **singleelectron_trigger_defaults,
-                        },
-                    ],
-                    "2016postVFP": [
-                        {
-                            "flagname": "trg_single_ele25",
-                            "hlt_path": "HLT_Ele25_eta2p1_WPTight_Gsf",
-                            "ptcut": 26,
-                            "etacut": 2.1,
-                            **singleelectron_trigger_defaults,
-                        },
-                    ],
-                    "2016preVFP": [
-                        {
-                            "flagname": "trg_single_ele25",
-                            "hlt_path": "HLT_Ele25_eta2p1_WPTight_Gsf",
-                            "ptcut": 26,
-                            "etacut": 2.1,
-                            **singleelectron_trigger_defaults,
-                        },
-                    ],
-                }
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_single_ele27",
+                                "hlt_path": "HLT_Ele27_WPTight_Gsf",
+                                "ptcut": 28,
+                                "etacut": 2.1,
+                                **singleelectron_trigger_defaults,
+                            },
+                            {
+                                "flagname": "trg_single_ele32",
+                                "hlt_path": "HLT_Ele32_WPTight_Gsf",
+                                "ptcut": 33,
+                                "etacut": 2.1,
+                                **singleelectron_trigger_defaults,
+                            },
+                            {
+                                "flagname": "trg_single_ele35",
+                                "hlt_path": "HLT_Ele35_WPTight_Gsf",
+                                "ptcut": 36,
+                                "etacut": 2.1,
+                                **singleelectron_trigger_defaults,
+                            },
+                        ]
+                        for era in ["2017", "2018"]
+                    },
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_single_ele25",
+                                "hlt_path": "HLT_Ele25_eta2p1_WPTight_Gsf",
+                                "ptcut": 26,
+                                "etacut": 2.1,
+                                **singleelectron_trigger_defaults,
+                            },
+                        ]
+                        for era in ["2016preVFP", "2016postVFP"]
+                    },
+                },
+                default=[  # 2022preEE, 2022postEE, 2023preBPix, 2023postBPix, 2024, 2025, 2026
+                    {
+                        "flagname": "trg_single_ele30",
+                        "hlt_path": "HLT_Ele30_WPTight_Gsf",
+                        "ptcut": 32,
+                        "etacut": 2.5,
+                        **singleelectron_trigger_defaults,
+                    },
+                ],
             ),
         },
     )
@@ -522,32 +395,22 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
         {
             "eltau_cross_trigger": EraModifier(
                 {
-                    "2026": [
-                        {
-                            "flagname": "trg_cross_ele24tau30_pnet",
-                            "hlt_path": "HLT_Ele24_eta2p1_WPTight_Gsf_PNetTauhPFJet30_Medium_eta2p3_CrossL1",
-                            "p1_ptcut": 25,
-                            "p2_ptcut": 35,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1,
-                            "p1_filterbit": 2,
-                            "p2_filterbit": "1, 4, 12",
-                            **electron_tau_cross_trigger_defaults,
-                        },
-                    ],
-                    "2025": [
-                        {
-                            "flagname": "trg_cross_ele24tau30_pnet",
-                            "hlt_path": "HLT_Ele24_eta2p1_WPTight_Gsf_PNetTauhPFJet30_Medium_eta2p3_CrossL1",
-                            "p1_ptcut": 25,
-                            "p2_ptcut": 35,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1,
-                            "p1_filterbit": 2,
-                            "p2_filterbit": "1, 4, 12",
-                            **electron_tau_cross_trigger_defaults,
-                        },
-                    ],
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_cross_ele24tau30_pnet",
+                                "hlt_path": "HLT_Ele24_eta2p1_WPTight_Gsf_PNetTauhPFJet30_Medium_eta2p3_CrossL1",
+                                "p1_ptcut": 25,
+                                "p2_ptcut": 35,
+                                "p1_etacut": 2.1,
+                                "p2_etacut": 2.1,
+                                "p1_filterbit": 2,
+                                "p2_filterbit": "1, 4, 12",
+                                **electron_tau_cross_trigger_defaults,
+                            },
+                        ]
+                        for era in ["2025", "2026"]
+                    },
                     "2024": [
                         {
                             "flagname": "trg_cross_ele24tau30_hps",
@@ -555,61 +418,9 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
                             "p1_ptcut": 25,
                             "p2_ptcut": 35,
                             "p1_etacut": 2.1,
-                            "p2_etacut": 2.1, 
-                            "p1_filterbit": 2, 
-                            "p2_filterbit": "3, 12",  
-                            **electron_tau_cross_trigger_defaults,
-                        },
-                    ],
-                    "2023postBPix": [
-                        {
-                            "flagname": "trg_cross_ele24tau30_hps",
-                            "hlt_path": "HLT_Ele24_eta2p1_WPTight_Gsf_LooseDeepTauPFTauHPS30_eta2p1_CrossL1",
-                            "p1_ptcut": 25,
-                            "p2_ptcut": 35,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1, 
-                            "p1_filterbit": 2, 
-                            "p2_filterbit": "3, 8",  
-                            **electron_tau_cross_trigger_defaults,
-                        },
-                    ],
-                    "2023preBPix": [
-                        {
-                            "flagname": "trg_cross_ele24tau30_hps",
-                            "hlt_path": "HLT_Ele24_eta2p1_WPTight_Gsf_LooseDeepTauPFTauHPS30_eta2p1_CrossL1",
-                            "p1_ptcut": 25,
-                            "p2_ptcut": 35,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1, 
-                            "p1_filterbit": 2, 
-                            "p2_filterbit": "3, 8",  
-                            **electron_tau_cross_trigger_defaults,
-                        },
-                    ],
-                    "2022postEE": [
-                        {
-                            "flagname": "trg_cross_ele24tau30_hps",
-                            "hlt_path": "HLT_Ele24_eta2p1_WPTight_Gsf_LooseDeepTauPFTauHPS30_eta2p1_CrossL1",
-                            "p1_ptcut": 25,
-                            "p2_ptcut": 35,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1, 
-                            "p1_filterbit": 2, 
-                            "p2_filterbit": "3, 8",  
-                            **electron_tau_cross_trigger_defaults,
-                        },
-                    ],
-                    "2022preEE": [
-                        {
-                            "flagname": "trg_cross_ele24tau30_hps",
-                            "hlt_path": "HLT_Ele24_eta2p1_WPTight_Gsf_LooseDeepTauPFTauHPS30_eta2p1_CrossL1",
-                            "p1_ptcut": 25,
-                            "p2_ptcut": 35,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1, 
-                            "p1_filterbit": 2, 
-                            "p2_filterbit": "3, 8",  
+                            "p2_etacut": 2.1,
+                            "p1_filterbit": 2,
+                            "p2_filterbit": "3, 12",
                             **electron_tau_cross_trigger_defaults,
                         },
                     ],
@@ -723,7 +534,20 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
                             **electron_tau_cross_trigger_defaults,
                         },
                     ],
-                }
+                },
+                default=[  # 2022preEE, 2022postEE, 2023preBPix, 2023postBPix
+                    {
+                        "flagname": "trg_cross_ele24tau30_hps",
+                        "hlt_path": "HLT_Ele24_eta2p1_WPTight_Gsf_LooseDeepTauPFTauHPS30_eta2p1_CrossL1",
+                        "p1_ptcut": 25,
+                        "p2_ptcut": 35,
+                        "p1_etacut": 2.1,
+                        "p2_etacut": 2.1,
+                        "p1_filterbit": 2,
+                        "p2_filterbit": "3, 8",
+                        **electron_tau_cross_trigger_defaults,
+                    },
+                ],
             ),
         },
     )
@@ -740,97 +564,22 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
         {
             "doubletau_trigger": EraModifier(
                 {
-                    "2026": [
-                        {
-                            "flagname": "trg_double_tau30_mediumiso_pnet",
-                            "hlt_path": "HLT_DoublePNetTauhPFJet30_Medium_L2NN_eta2p3",
-                            "p1_ptcut": 35,
-                            "p2_ptcut": 35,
-                            "p1_etacut": 2.3,
-                            "p2_etacut": 2.3,
-                            "p1_filterbit": "1, 4, 11",
-                            "p2_filterbit": "1, 4, 11",
-                            **doubletau_trigger_defaults,
-                        },
-                    ],
-                    "2025": [
-                        {
-                            "flagname": "trg_double_tau30_mediumiso_pnet",
-                            "hlt_path": "HLT_DoublePNetTauhPFJet30_Medium_L2NN_eta2p3",
-                            "p1_ptcut": 35,
-                            "p2_ptcut": 35,
-                            "p1_etacut": 2.3,
-                            "p2_etacut": 2.3,
-                            "p1_filterbit": "1, 4, 11",
-                            "p2_filterbit": "1, 4, 11",
-                            **doubletau_trigger_defaults,
-                        },
-                    ],
-                    "2024": [
-                        {
-                            "flagname": "trg_double_tau30_mediumiso_pnet",
-                            "hlt_path": "HLT_DoublePNetTauhPFJet30_Medium_L2NN_eta2p3",
-                            "p1_ptcut": 35,
-                            "p2_ptcut": 35,
-                            "p1_etacut": 2.3,
-                            "p2_etacut": 2.3,
-                            "p1_filterbit": "1, 4, 11",
-                            "p2_filterbit": "1, 4, 11",
-                            **doubletau_trigger_defaults,
-                        },
-                    ],
-                    "2023postBPix": [
-                        {
-                            "flagname": "trg_double_tau35_mediumiso_hps",
-                            "hlt_path": "HLT_DoubleMediumDeepTauPFTauHPS35_L2NN_eta2p1",
-                            "p1_ptcut": 40,
-                            "p2_ptcut": 40,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1,
-                            "p1_filterbit": "3, 7",
-                            "p2_filterbit": "3, 7",
-                            **doubletau_trigger_defaults,
-                        },
-                    ],
-                    "2023preBPix": [
-                        {
-                            "flagname": "trg_double_tau35_mediumiso_hps",
-                            "hlt_path": "HLT_DoubleMediumDeepTauPFTauHPS35_L2NN_eta2p1",
-                            "p1_ptcut": 40,
-                            "p2_ptcut": 40,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1,
-                            "p1_filterbit": "3, 7",
-                            "p2_filterbit": "3, 7",
-                            **doubletau_trigger_defaults,
-                        },
-                    ],
-                    "2022postEE": [
-                        {
-                            "flagname": "trg_double_tau35_mediumiso_hps",
-                            "hlt_path": "HLT_DoubleMediumDeepTauPFTauHPS35_L2NN_eta2p1",
-                            "p1_ptcut": 40,
-                            "p2_ptcut": 40,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1,
-                            "p1_filterbit": "3, 7",
-                            "p2_filterbit": "3, 7",
-                            **doubletau_trigger_defaults,
-                        },
-                    ],
-                    "2022preEE": [
-                        {
-                            "flagname": "trg_double_tau35_mediumiso_hps",
-                            "hlt_path": "HLT_DoubleMediumDeepTauPFTauHPS35_L2NN_eta2p1",
-                            "p1_ptcut": 40,
-                            "p2_ptcut": 40,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1,
-                            "p1_filterbit": "3, 7",
-                            "p2_filterbit": "3, 7",
-                            **doubletau_trigger_defaults,
-                        },
-                    ],
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_double_tau30_mediumiso_pnet",
+                                "hlt_path": "HLT_DoublePNetTauhPFJet30_Medium_L2NN_eta2p3",
+                                "p1_ptcut": 35,
+                                "p2_ptcut": 35,
+                                "p1_etacut": 2.3,
+                                "p2_etacut": 2.3,
+                                "p1_filterbit": "1, 4, 11",
+                                "p2_filterbit": "1, 4, 11",
+                                **doubletau_trigger_defaults,
+                            },
+                        ]
+                        for era in ["2024", "2025", "2026"]
+                    },
                     "2018": [
                         {
                             "flagname": "trg_double_tau35_mediumiso_hps",
@@ -913,55 +662,47 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
                             **doubletau_trigger_defaults,
                         },
                     ],
-                    "2016preVFP": [
-                        {
-                            "flagname": "trg_double_tau35_mediumiso",
-                            "hlt_path": "HLT_DoubleMediumIsoPFTau35_Trk1_eta2p1_Reg",
-                            "p1_ptcut": 40,
-                            "p2_ptcut": 40,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1,
-                            "p1_filterbit": -1,  # TODO switch to "p1_filterbit": 6, if the bits are correct
-                            "p2_filterbit": -1,  # TODO switch to "p2_filterbit": 6, if the bits are correct
-                            **doubletau_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_double_tau35_mediumcombiso",
-                            "hlt_path": "HLT_DoubleMediumCombinedIsoPFTau35_Trk1_eta2p1_Reg",
-                            "p1_ptcut": 40,
-                            "p2_ptcut": 40,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1,
-                            "p1_filterbit": -1,  # TODO switch to "p1_filterbit": 6, if the bits are correct
-                            "p2_filterbit": -1,  # TODO switch to "p2_filterbit": 6, if the bits are correct
-                            **doubletau_trigger_defaults,
-                        },
-                    ],
-                    "2016postVFP": [
-                        {
-                            "flagname": "trg_double_tau35_mediumiso",
-                            "hlt_path": "HLT_DoubleMediumIsoPFTau35_Trk1_eta2p1_Reg",
-                            "p1_ptcut": 40,
-                            "p2_ptcut": 40,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1,
-                            "p1_filterbit": -1,  # TODO switch to "p1_filterbit": 6, if the bits are correct
-                            "p2_filterbit": -1,  # TODO switch to "p2_filterbit": 6, if the bits are correct
-                            **doubletau_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_double_tau35_mediumcombiso",
-                            "hlt_path": "HLT_DoubleMediumCombinedIsoPFTau35_Trk1_eta2p1_Reg",
-                            "p1_ptcut": 40,
-                            "p2_ptcut": 40,
-                            "p1_etacut": 2.1,
-                            "p2_etacut": 2.1,
-                            "p1_filterbit": -1,  # TODO switch to "p1_filterbit": 6, if the bits are correct
-                            "p2_filterbit": -1,  # TODO switch to "p2_filterbit": 6, if the bits are correct
-                            **doubletau_trigger_defaults,
-                        },
-                    ],
-                }
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_double_tau35_mediumiso",
+                                "hlt_path": "HLT_DoubleMediumIsoPFTau35_Trk1_eta2p1_Reg",
+                                "p1_ptcut": 40,
+                                "p2_ptcut": 40,
+                                "p1_etacut": 2.1,
+                                "p2_etacut": 2.1,
+                                "p1_filterbit": -1,  # TODO switch to "p1_filterbit": 6, if the bits are correct
+                                "p2_filterbit": -1,  # TODO switch to "p2_filterbit": 6, if the bits are correct
+                                **doubletau_trigger_defaults,
+                            },
+                            {
+                                "flagname": "trg_double_tau35_mediumcombiso",
+                                "hlt_path": "HLT_DoubleMediumCombinedIsoPFTau35_Trk1_eta2p1_Reg",
+                                "p1_ptcut": 40,
+                                "p2_ptcut": 40,
+                                "p1_etacut": 2.1,
+                                "p2_etacut": 2.1,
+                                "p1_filterbit": -1,  # TODO switch to "p1_filterbit": 6, if the bits are correct
+                                "p2_filterbit": -1,  # TODO switch to "p2_filterbit": 6, if the bits are correct
+                                **doubletau_trigger_defaults,
+                            },
+                        ]
+                        for era in ["2016preVFP", "2016postVFP"]
+                    },
+                },
+                default=[  # 2022preEE, 2022postEE, 2023preBPix, 2023postBPix
+                    {
+                        "flagname": "trg_double_tau35_mediumiso_hps",
+                        "hlt_path": "HLT_DoubleMediumDeepTauPFTauHPS35_L2NN_eta2p1",
+                        "p1_ptcut": 40,
+                        "p2_ptcut": 40,
+                        "p1_etacut": 2.1,
+                        "p2_etacut": 2.1,
+                        "p1_filterbit": "3, 7",
+                        "p2_filterbit": "3, 7",
+                        **doubletau_trigger_defaults,
+                    },
+                ],
             ),
         },
     )
@@ -982,69 +723,6 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
         {
             "elmu_cross_trigger": EraModifier(
                 {
-                    "2026": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": "trg_cross_mu23ele12",
-                            "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL_DZ",
-                            "p1_ptcut": 12,
-                            "p2_ptcut": 24,
-                            **elmu_cross_trigger_defaults,
-                        },
-                    ],
-                    "2025": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": "trg_cross_mu23ele12",
-                            "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL_DZ",
-                            "p1_ptcut": 12,
-                            "p2_ptcut": 24,
-                            **elmu_cross_trigger_defaults,
-                        },
-                    ],
-                    "2024": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": "trg_cross_mu23ele12",
-                            "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL_DZ",
-                            "p1_ptcut": 12,
-                            "p2_ptcut": 24,
-                            **elmu_cross_trigger_defaults,
-                        },
-                    ],
-                    "2023postBPix": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": "trg_cross_mu23ele12",
-                            "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL_DZ",
-                            "p1_ptcut": 12,
-                            "p2_ptcut": 24,
-                            **elmu_cross_trigger_defaults,
-                        },
-                    ],
-                    "2023preBPix": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": "trg_cross_mu23ele12",
-                            "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL_DZ",
-                            "p1_ptcut": 12,
-                            "p2_ptcut": 24,
-                            **elmu_cross_trigger_defaults,
-                        },
-                    ],
-                    "2022postEE": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": "trg_cross_mu23ele12",
-                            "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL_DZ",
-                            "p1_ptcut": 12,
-                            "p2_ptcut": 24,
-                            **elmu_cross_trigger_defaults,
-                        },
-                    ],
-                    "2022preEE": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": "trg_cross_mu23ele12",
-                            "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL_DZ",
-                            "p1_ptcut": 12,
-                            "p2_ptcut": 24,
-                            **elmu_cross_trigger_defaults,
-                        },
-                    ],
                     "2018": [
                         {
                             "flagname": "trg_cross_mu23ele12",
@@ -1077,67 +755,49 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
                             **elmu_cross_trigger_defaults,
                         },
                     ],
-                    "2016postVFP": [
-                        {
-                            "flagname": "trg_cross_mu23ele12_dz",
-                            "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL_DZ",
-                            "p1_ptcut": 13,
-                            "p2_ptcut": 24,
-                            **elmu_cross_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_cross_mu8ele23_dz",
-                            "hlt_path": "HLT_Mu8_TrkIsoVVL_Ele23_CaloIdL_TrackIdL_IsoVL_DZ",
-                            "p1_ptcut": 24,
-                            "p2_ptcut": 9,
-                            **elmu_cross_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_cross_mu23ele12",
-                            "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL",
-                            "p1_ptcut": 13,
-                            "p2_ptcut": 24,
-                            **elmu_cross_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_cross_mu8ele23",
-                            "hlt_path": "HLT_Mu8_TrkIsoVVL_Ele23_CaloIdL_TrackIdL_IsoVL",
-                            "p1_ptcut": 24,
-                            "p2_ptcut": 9,
-                            **elmu_cross_trigger_defaults,
-                        },
-                    ],
-                    "2016preVFP": [
-                        {
-                            "flagname": "trg_cross_mu23ele12_dz",
-                            "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL_DZ",
-                            "p1_ptcut": 13,
-                            "p2_ptcut": 24,
-                            **elmu_cross_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_cross_mu8ele23_dz",
-                            "hlt_path": "HLT_Mu8_TrkIsoVVL_Ele23_CaloIdL_TrackIdL_IsoVL_DZ",
-                            "p1_ptcut": 24,
-                            "p2_ptcut": 9,
-                            **elmu_cross_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_cross_mu23ele12",
-                            "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL",
-                            "p1_ptcut": 13,
-                            "p2_ptcut": 24,
-                            **elmu_cross_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_cross_mu8ele23",
-                            "hlt_path": "HLT_Mu8_TrkIsoVVL_Ele23_CaloIdL_TrackIdL_IsoVL",
-                            "p1_ptcut": 24,
-                            "p2_ptcut": 9,
-                            **elmu_cross_trigger_defaults,
-                        },
-                    ],
-                }
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_cross_mu23ele12_dz",
+                                "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL_DZ",
+                                "p1_ptcut": 13,
+                                "p2_ptcut": 24,
+                                **elmu_cross_trigger_defaults,
+                            },
+                            {
+                                "flagname": "trg_cross_mu8ele23_dz",
+                                "hlt_path": "HLT_Mu8_TrkIsoVVL_Ele23_CaloIdL_TrackIdL_IsoVL_DZ",
+                                "p1_ptcut": 24,
+                                "p2_ptcut": 9,
+                                **elmu_cross_trigger_defaults,
+                            },
+                            {
+                                "flagname": "trg_cross_mu23ele12",
+                                "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL",
+                                "p1_ptcut": 13,
+                                "p2_ptcut": 24,
+                                **elmu_cross_trigger_defaults,
+                            },
+                            {
+                                "flagname": "trg_cross_mu8ele23",
+                                "hlt_path": "HLT_Mu8_TrkIsoVVL_Ele23_CaloIdL_TrackIdL_IsoVL",
+                                "p1_ptcut": 24,
+                                "p2_ptcut": 9,
+                                **elmu_cross_trigger_defaults,
+                            },
+                        ]
+                        for era in ["2016postVFP", "2016preVFP"]
+                    },
+                },
+                default=[  ## TODO: not implemented, here as a placeholder (2022preEE, 2022postEE, 2023preBPix, 2023postBPix, 2024, 2025, 2026)
+                    {
+                        "flagname": "trg_cross_mu23ele12",
+                        "hlt_path": "HLT_Mu23_TrkIsoVVL_Ele12_CaloIdL_TrackIdL_IsoVL_DZ",
+                        "p1_ptcut": 12,
+                        "p2_ptcut": 24,
+                        **elmu_cross_trigger_defaults,
+                    },
+                ],
             ),
         },
     )
@@ -1155,24 +815,18 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
         {
             "singletau_trigger_leading": EraModifier(
                 {
-                    "2026": [
-                        {
-                            "flagname": "trg_single_tau130_pnet_1",
-                            "hlt_path": "HLT_SinglePNetTauhPFJet130_Medium_L2NN_eta2p3",
-                            "ptcut": 130,
-                            "filterbit": "1, 4, 9",
-                            **singletau_trigger_defaults,
-                        },
-                    ],
-                    "2025": [
-                        {
-                            "flagname": "trg_single_tau130_pnet_1",
-                            "hlt_path": "HLT_SinglePNetTauhPFJet130_Medium_L2NN_eta2p3",
-                            "ptcut": 130,
-                            "filterbit": "1, 4, 9",
-                            **singletau_trigger_defaults,
-                        },
-                    ],
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_single_tau130_pnet_1",
+                                "hlt_path": "HLT_SinglePNetTauhPFJet130_Medium_L2NN_eta2p3",
+                                "ptcut": 130,
+                                "filterbit": "1, 4, 9",
+                                **singletau_trigger_defaults,
+                            },
+                        ]
+                        for era in ["2025", "2026"]
+                    },
                     "2024": [
                         {
                             "flagname": "trg_single_tau180_hps_1",
@@ -1182,87 +836,44 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
                             **singletau_trigger_defaults,
                         },
                     ],
-                    "2023postBPix": [
-                        {
-                            "flagname": "trg_single_tau180_hps_1",
-                            "hlt_path": "HLT_LooseDeepTauPFTauHPS180_L2NN_eta2p1",
-                            "ptcut": 180,
-                            "filterbit": "3, 10",
-                            **singletau_trigger_defaults,
-                        },
-                    ],
-                    "2023preBPix": [
-                        {
-                            "flagname": "trg_single_tau180_hps_1",
-                            "hlt_path": "HLT_LooseDeepTauPFTauHPS180_L2NN_eta2p1",
-                            "ptcut": 180,
-                            "filterbit": "3, 10",
-                            **singletau_trigger_defaults,
-                        },
-                    ],
-                    "2022postEE": [
-                        {
-                            "flagname": "trg_single_tau180_hps_1",
-                            "hlt_path": "HLT_LooseDeepTauPFTauHPS180_L2NN_eta2p1",
-                            "ptcut": 180,
-                            "filterbit": "3, 10",
-                            **singletau_trigger_defaults,
-                        },
-                    ],
-                    "2022preEE": [
-                        {
-                            "flagname": "trg_single_tau180_hps_1",
-                            "hlt_path": "HLT_LooseDeepTauPFTauHPS180_L2NN_eta2p1",
-                            "ptcut": 180,
-                            "filterbit": "3, 10",
-                            **singletau_trigger_defaults,
-                        },
-                    ],
-                    "2018": [
-                        {
-                            "flagname": "trg_single_tau180_1",
-                            "hlt_path": "HLT_MediumChargedIsoPFTau180HighPtRelaxedIso_Trk50_eta2p1",
-                            "ptcut": 180,
-                            **singletau_trigger_defaults,
-                        }
-                    ],
-                    "2017": [
-                        {
-                            "flagname": "trg_single_tau180_1",
-                            "hlt_path": "HLT_MediumChargedIsoPFTau180HighPtRelaxedIso_Trk50_eta2p1",
-                            "ptcut": 180,
-                            **singletau_trigger_defaults,
-                        }
-                    ],
-                    "2016postVFP": [
-                        {
-                            "flagname": "trg_single_tau120_1",
-                            "hlt_path": "HLT_VLooseIsoPFTau120_Trk50_eta2p1",
-                            "ptcut": 120,
-                            **singletau_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_single_tau140_1",
-                            "hlt_path": "HLT_VLooseIsoPFTau140_Trk50_eta2p1",
-                            "ptcut": 140,
-                            **singletau_trigger_defaults,
-                        },
-                    ],
-                    "2016preVFP": [
-                        {
-                            "flagname": "trg_single_tau120_1",
-                            "hlt_path": "HLT_VLooseIsoPFTau120_Trk50_eta2p1",
-                            "ptcut": 120,
-                            **singletau_trigger_defaults,
-                        },
-                        {
-                            "flagname": "trg_single_tau140_1",
-                            "hlt_path": "HLT_VLooseIsoPFTau140_Trk50_eta2p1",
-                            "ptcut": 140,
-                            **singletau_trigger_defaults,
-                        },
-                    ],
-                }
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_single_tau180_1",
+                                "hlt_path": "HLT_MediumChargedIsoPFTau180HighPtRelaxedIso_Trk50_eta2p1",
+                                "ptcut": 180,
+                                **singletau_trigger_defaults,
+                            }
+                        ]
+                        for era in ["2017", "2018"]
+                    },
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_single_tau120_1",
+                                "hlt_path": "HLT_VLooseIsoPFTau120_Trk50_eta2p1",
+                                "ptcut": 120,
+                                **singletau_trigger_defaults,
+                            },
+                            {
+                                "flagname": "trg_single_tau140_1",
+                                "hlt_path": "HLT_VLooseIsoPFTau140_Trk50_eta2p1",
+                                "ptcut": 140,
+                                **singletau_trigger_defaults,
+                            },
+                        ]
+                        for era in ["2016preVFP", "2016postVFP"]
+                    },
+                },
+                default=[  # 2022preEE, 2022postEE, 2023preBPix, 2023postBPix
+                    {
+                        "flagname": "trg_single_tau180_hps_1",
+                        "hlt_path": "HLT_LooseDeepTauPFTauHPS180_L2NN_eta2p1",
+                        "ptcut": 180,
+                        "filterbit": "3, 10",
+                        **singletau_trigger_defaults,
+                    },
+                ],
             )
         },
     )
@@ -1280,24 +891,18 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
         {
             "singletau_trigger_trailing": EraModifier(
                 {
-                    "2026": [
-                        {
-                            "flagname": "trg_single_tau130_pnet_2",
-                            "hlt_path": "HLT_SinglePNetTauhPFJet130_Medium_L2NN_eta2p3",
-                            "ptcut": 130,
-                            "filterbit": "1, 4, 9",
-                            **singletau_trigger_trailing_defaults,
-                        },
-                    ],
-                    "2025": [
-                        {
-                            "flagname": "trg_single_tau130_pnet_2",
-                            "hlt_path": "HLT_SinglePNetTauhPFJet130_Medium_L2NN_eta2p3",
-                            "ptcut": 130,
-                            "filterbit": "1, 4, 9",
-                            **singletau_trigger_trailing_defaults,
-                        },
-                    ],
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_single_tau130_pnet_2",
+                                "hlt_path": "HLT_SinglePNetTauhPFJet130_Medium_L2NN_eta2p3",
+                                "ptcut": 130,
+                                "filterbit": "1, 4, 9",
+                                **singletau_trigger_trailing_defaults,
+                            },
+                        ]
+                        for era in ["2025", "2026"]
+                    },
                     "2024": [
                         {
                             "flagname": "trg_single_tau180_hps_2",
@@ -1307,87 +912,44 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
                             **singletau_trigger_trailing_defaults,
                         },
                     ],
-                    "2023postBPix": [
-                        {
-                            "flagname": "trg_single_tau180_hps_2",
-                            "hlt_path": "HLT_LooseDeepTauPFTauHPS180_L2NN_eta2p1",
-                            "ptcut": 180,
-                            "filterbit": "3, 10",
-                            **singletau_trigger_trailing_defaults,
-                        },
-                    ],
-                    "2023preBPix": [
-                        {
-                            "flagname": "trg_single_tau180_hps_2",
-                            "hlt_path": "HLT_LooseDeepTauPFTauHPS180_L2NN_eta2p1",
-                            "ptcut": 180,
-                            "filterbit": "3, 10",
-                            **singletau_trigger_trailing_defaults,
-                        },
-                    ],
-                    "2022postEE": [
-                        {
-                            "flagname": "trg_single_tau180_hps_2",
-                            "hlt_path": "HLT_LooseDeepTauPFTauHPS180_L2NN_eta2p1",
-                            "ptcut": 180,
-                            "filterbit": "3, 10",
-                            **singletau_trigger_trailing_defaults,
-                        },
-                    ],
-                    "2022preEE": [
-                        {
-                            "flagname": "trg_single_tau180_hps_2",
-                            "hlt_path": "HLT_LooseDeepTauPFTauHPS180_L2NN_eta2p1",
-                            "ptcut": 180,
-                            "filterbit": "3, 10",
-                            **singletau_trigger_trailing_defaults,
-                        },
-                    ],
-                    "2018": [
-                        {
-                            "flagname": "trg_single_tau180_2",
-                            "hlt_path": "HLT_MediumChargedIsoPFTau180HighPtRelaxedIso_Trk50_eta2p1",
-                            "ptcut": 180,
-                            **singletau_trigger_trailing_defaults,
-                        }
-                    ],
-                    "2017": [
-                        {
-                            "flagname": "trg_single_tau180_2",
-                            "hlt_path": "HLT_MediumChargedIsoPFTau180HighPtRelaxedIso_Trk50_eta2p1",
-                            "ptcut": 180,
-                            **singletau_trigger_trailing_defaults,
-                        }
-                    ],
-                    "2016postVFP": [
-                        {
-                            "flagname": "trg_single_tau120_2",
-                            "hlt_path": "HLT_VLooseIsoPFTau120_Trk50_eta2p1_v",
-                            "ptcut": 120,
-                            **singletau_trigger_trailing_defaults,
-                        },
-                        {
-                            "flagname": "trg_single_tau140_2",
-                            "hlt_path": "HLT_VLooseIsoPFTau140_Trk50_eta2p1_v",
-                            "ptcut": 140,
-                            **singletau_trigger_trailing_defaults,
-                        },
-                    ],
-                    "2016preVFP": [
-                        {
-                            "flagname": "trg_single_tau120_2",
-                            "hlt_path": "HLT_VLooseIsoPFTau120_Trk50_eta2p1_v",
-                            "ptcut": 120,
-                            **singletau_trigger_trailing_defaults,
-                        },
-                        {
-                            "flagname": "trg_single_tau140_2",
-                            "hlt_path": "HLT_VLooseIsoPFTau140_Trk50_eta2p1_v",
-                            "ptcut": 140,
-                            **singletau_trigger_trailing_defaults,
-                        },
-                    ],
-                }
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_single_tau180_2",
+                                "hlt_path": "HLT_MediumChargedIsoPFTau180HighPtRelaxedIso_Trk50_eta2p1",
+                                "ptcut": 180,
+                                **singletau_trigger_trailing_defaults,
+                            }
+                        ]
+                        for era in ["2017", "2018"]
+                    },
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_single_tau120_2",
+                                "hlt_path": "HLT_VLooseIsoPFTau120_Trk50_eta2p1_v",
+                                "ptcut": 120,
+                                **singletau_trigger_trailing_defaults,
+                            },
+                            {
+                                "flagname": "trg_single_tau140_2",
+                                "hlt_path": "HLT_VLooseIsoPFTau140_Trk50_eta2p1_v",
+                                "ptcut": 140,
+                                **singletau_trigger_trailing_defaults,
+                            },
+                        ]
+                        for era in ["2016preVFP", "2016postVFP"]
+                    },
+                },
+                default=[  # 2022preEE, 2022postEE, 2023preBPix, 2023postBPix
+                    {
+                        "flagname": "trg_single_tau180_hps_2",
+                        "hlt_path": "HLT_LooseDeepTauPFTauHPS180_L2NN_eta2p1",
+                        "ptcut": 180,
+                        "filterbit": "3, 10",
+                        **singletau_trigger_trailing_defaults,
+                    },
+                ],
             )
         },
     )
@@ -1441,37 +1003,20 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
 
             "mutau_cross_trigger_leg1_sf_file": EraModifier(
                 {
-                    "2016preVFP": '""',
-                    "2016postVFP": '""', 
-                    "2017": '""',
-                    "2018": '""',
                     "2022preEE": "data/hleprare/TriggerScaleFactors/2022preEE/CrossMuTauHlt_MuLeg_v1.json",
                     "2022postEE": "data/hleprare/TriggerScaleFactors/2022postEE/CrossMuTauHlt_MuLeg_v1.json",
                     "2023preBPix": "data/hleprare/TriggerScaleFactors/2023preBPix/CrossMuTauHlt_MuLeg_v1.json",
-                    "2023postBPix": "data/hleprare/TriggerScaleFactors/2023postBPix/CrossMuTauHlt_MuLeg_v1.json",
-                    "2024":"data/hleprare/TriggerScaleFactors/2023postBPix/CrossMuTauHlt_MuLeg_v1.json",
-                    "2025":"data/hleprare/TriggerScaleFactors/2023postBPix/CrossMuTauHlt_MuLeg_v1.json",
-                    "2026":"data/hleprare/TriggerScaleFactors/2023postBPix/CrossMuTauHlt_MuLeg_v1.json",
-                }
+                    **{
+                        era: "data/hleprare/TriggerScaleFactors/2023postBPix/CrossMuTauHlt_MuLeg_v1.json"
+                        for era in ["2023postBPix", "2024", "2025", "2026"]
+                    },
+                },
+                default='""',  # Run 2 (unused)
             ),
             "mutau_trigger_leg1_sf": [
                 {
                     "mutau_cross_trigger_leg1_flagname": "trg_wgt_mu20tau27_leg1",
-                    "mutau_cross_trigger_flag": EraModifier(
-                        {
-                            "2016preVFP": "trg_cross_mu20tau27_hps",
-                            "2016postVFP": "trg_cross_mu20tau27_hps",
-                            "2017": "trg_cross_mu20tau27_hps",
-                            "2018": "trg_cross_mu20tau27_hps",
-                            "2022preEE": "trg_cross_mu20tau27_hps",
-                            "2022postEE": "trg_cross_mu20tau27_hps",
-                            "2023preBPix": "trg_cross_mu20tau27_hps",
-                            "2023postBPix": "trg_cross_mu20tau27_hps",
-                            "2024": "trg_cross_mu20tau27_hps",
-                            "2025": "trg_cross_mu20tau27_pnet",
-                            "2026": "trg_cross_mu20tau27_pnet",
-                        }
-                    ),
+                    "mutau_cross_trigger_flag": MUTAU_CROSS_TRIGGER_FLAG,
                     "mutau_cross_trigger_leg1_sf_name": "NUM_IsoMu20_DEN_CutBasedIdTight_and_PFIsoTight",
                     "mutau_cross_trigger_leg1_variation": "nominal",
                 },
@@ -1479,21 +1024,7 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
             "mutau_trigger_leg2_sf": [
                 {
                     "mutau_cross_trigger_leg2_flagname": "trg_wgt_mu20tau27_leg2",
-                    "mutau_cross_trigger_flag": EraModifier(
-                        {
-                            "2016preVFP": "trg_cross_mu20tau27_hps",
-                            "2016postVFP": "trg_cross_mu20tau27_hps",
-                            "2017": "trg_cross_mu20tau27_hps",
-                            "2018": "trg_cross_mu20tau27_hps",
-                            "2022preEE": "trg_cross_mu20tau27_hps",
-                            "2022postEE": "trg_cross_mu20tau27_hps",
-                            "2023preBPix": "trg_cross_mu20tau27_hps",
-                            "2023postBPix": "trg_cross_mu20tau27_hps",
-                            "2024": "trg_cross_mu20tau27_hps",
-                            "2025": "trg_cross_mu20tau27_pnet",
-                            "2026": "trg_cross_mu20tau27_pnet",
-                        }
-                    ),
+                    "mutau_cross_trigger_flag": MUTAU_CROSS_TRIGGER_FLAG,
                     "mutau_cross_trigger_leg2_sf_name": "mutau",
                     "mutau_cross_trigger_leg2_variation": "nom",
                 },
@@ -1507,33 +1038,26 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
         {
             "singleelectron_trigger_sf_file": EraModifier(
                 {
-                    "2016preVFP": '""',
-                    "2016postVFP": '""', 
-                    "2017": '""',
-                    "2018": '""',
                     "2022preEE": "/cvmfs/cms-griddata.cern.ch/cat/metadata/EGM/Run3-22CDSep23-Summer22-NanoAODv12/2025-12-15/electronHlt.json.gz",
                     "2022postEE": "/cvmfs/cms-griddata.cern.ch/cat/metadata/EGM/Run3-22EFGSep23-Summer22EE-NanoAODv12/2025-12-15/electronHlt.json.gz",
                     "2023preBPix": "/cvmfs/cms-griddata.cern.ch/cat/metadata/EGM/Run3-23CSep23-Summer23-NanoAODv12/2025-12-15/electronHlt.json.gz",
                     "2023postBPix": "/cvmfs/cms-griddata.cern.ch/cat/metadata/EGM/Run3-23DSep23-Summer23BPix-NanoAODv12/2025-12-15/electronHlt.json.gz",
-                    "2024": "/cvmfs/cms-griddata.cern.ch/cat/metadata/EGM/Run3-24CDEReprocessingFGHIPrompt-Summer24-NanoAODv15/2025-12-15/electronHlt.json.gz",
-                    "2025": "/cvmfs/cms-griddata.cern.ch/cat/metadata/EGM/Run3-24CDEReprocessingFGHIPrompt-Summer24-NanoAODv15/2025-12-15/electronHlt.json.gz",
-                    "2026": "/cvmfs/cms-griddata.cern.ch/cat/metadata/EGM/Run3-24CDEReprocessingFGHIPrompt-Summer24-NanoAODv15/2025-12-15/electronHlt.json.gz",
-                }
+                    **{
+                        era: "/cvmfs/cms-griddata.cern.ch/cat/metadata/EGM/Run3-24CDEReprocessingFGHIPrompt-Summer24-NanoAODv15/2025-12-15/electronHlt.json.gz"
+                        for era in ["2024", "2025", "2026"]
+                    },
+                },
+                default='""',  # Run 2 (unused)
             ),
             "singleelctron_trigger_era": EraModifier(
                 {
-                    "2016preVFP": '""',
-                    "2016postVFP": '""', 
-                    "2017": '""',
-                    "2018": '""',
                     "2022preEE": "2022Re-recoBCD",
                     "2022postEE": "2022Re-recoE+PromptFG",
                     "2023preBPix": "2023PromptC",
                     "2023postBPix": "2023PromptD",
-                    "2024": "2024Prompt",
-                    "2025": "2024Prompt",
-                    "2026": "2024Prompt",
-                }
+                    **{era: "2024Prompt" for era in ["2024", "2025", "2026"]},
+                },
+                default='""',  # Run 2 (unused)
             ),
             "singleelectron_trigger_sf": [
                 {
@@ -1547,36 +1071,19 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
 
             "eletau_cross_trigger_leg1_sf_file": EraModifier(
                 {
-                    "2016preVFP": '""',
-                    "2016postVFP": '""', 
-                    "2017": '""',
-                    "2018": '""',
                     "2022preEE": "data/hleprare/TriggerScaleFactors/2022preEE/CrossEleTauHlt_EleLeg_v1.json",
                     "2022postEE": "data/hleprare/TriggerScaleFactors/2022postEE/CrossEleTauHlt_EleLeg_v1.json",
                     "2023preBPix": "data/hleprare/TriggerScaleFactors/2023preBPix/CrossEleTauHlt_EleLeg_v1.json",
-                    "2023postBPix": "data/hleprare/TriggerScaleFactors/2023postBPix/CrossEleTauHlt_EleLeg_v1.json",
-                    "2024": "data/hleprare/TriggerScaleFactors/2023postBPix/CrossEleTauHlt_EleLeg_v1.json",
-                    "2025": "data/hleprare/TriggerScaleFactors/2023postBPix/CrossEleTauHlt_EleLeg_v1.json",
-                    "2026": "data/hleprare/TriggerScaleFactors/2023postBPix/CrossEleTauHlt_EleLeg_v1.json",
-                }
+                    **{
+                        era: "data/hleprare/TriggerScaleFactors/2023postBPix/CrossEleTauHlt_EleLeg_v1.json"
+                        for era in ["2023postBPix", "2024", "2025", "2026"]
+                    },
+                },
+                default='""',  # Run 2 (unused)
             ),
             "eletau_cross_trigger_leg1_sf": [
                 {
-                    "eletau_cross_trigger_flag": EraModifier(
-                        {
-                            "2016preVFP": "trg_cross_ele24tau30_hps",
-                            "2016postVFP": "trg_cross_ele24tau30_hps",
-                            "2017": "trg_cross_ele24tau30_hps",
-                            "2018": "trg_cross_ele24tau30_hps",
-                            "2022preEE": "trg_cross_ele24tau30_hps",
-                            "2022postEE": "trg_cross_ele24tau30_hps",
-                            "2023preBPix": "trg_cross_ele24tau30_hps",
-                            "2023postBPix": "trg_cross_ele24tau30_hps",
-                            "2024": "trg_cross_ele24tau30_hps",
-                            "2025": "trg_cross_ele24tau30_pnet",
-                            "2026": "trg_cross_ele24tau30_pnet",
-                        }
-                    ),
+                    "eletau_cross_trigger_flag": ELETAU_CROSS_TRIGGER_FLAG,
                     "eletau_cross_trigger_leg1_flagname": "trg_wgt_ele24tau30_leg1",
                     "eletau_cross_trigger_leg1_sf_name": "Electron-HLT-SF",
                     "eletau_cross_trigger_leg1_path_id_name": "HLT_SF_Ele24_TightID",
@@ -1586,21 +1093,7 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
             "eletau_cross_trigger_leg2_sf": [
                 {
                     "eletau_cross_trigger_leg2_flagname": "trg_wgt_ele24tau30_leg2",
-                    "eletau_cross_trigger_flag": EraModifier(
-                        {
-                            "2016preVFP": "trg_cross_ele24tau30_hps",
-                            "2016postVFP": "trg_cross_ele24tau30_hps",
-                            "2017": "trg_cross_ele24tau30_hps",
-                            "2018": "trg_cross_ele24tau30_hps",
-                            "2022preEE": "trg_cross_ele24tau30_hps",
-                            "2022postEE": "trg_cross_ele24tau30_hps",
-                            "2023preBPix": "trg_cross_ele24tau30_hps",
-                            "2023postBPix": "trg_cross_ele24tau30_hps",
-                            "2024": "trg_cross_ele24tau30_hps",
-                            "2025": "trg_cross_ele24tau30_pnet",
-                            "2026": "trg_cross_ele24tau30_pnet",
-                        }
-                    ),
+                    "eletau_cross_trigger_flag": ELETAU_CROSS_TRIGGER_FLAG,
                     "eletau_cross_trigger_leg2_sf_name": "etau",
                     "eletau_cross_trigger_leg2_variation": "nom",
                 },
@@ -1614,385 +1107,22 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
         {
             "doubletau_trigger_leg1_sf": [
                 {
-                    "doubletau_trigger_leg1_flagname": EraModifier(
-                        {
-                            "2016preVFP": '""',
-                            "2016postVFP": '""', 
-                            "2017": '""',
-                            "2018": '""',
-                            "2022preEE": "trg_wgt_doubletau35_leg1",
-                            "2022postEE": "trg_wgt_doubletau35_leg1",
-                            "2023preBPix": "trg_wgt_doubletau35_leg1",
-                            "2023postBPix": "trg_wgt_doubletau35_leg1",
-                            "2024": "trg_wgt_doubletau30_leg1",
-                            "2025": "trg_wgt_doubletau30_leg1",
-                            "2026": "trg_wgt_doubletau30_leg1",
-                        }
-                    ),
-                    "doubletau_trigger_flag": EraModifier(
-                        {
-                            "2016preVFP": '""',
-                            "2016postVFP": '""', 
-                            "2017": '""',
-                            "2018": '""',
-                            "2022preEE": "trg_double_tau35_mediumiso_hps",
-                            "2022postEE": "trg_double_tau35_mediumiso_hps",
-                            "2023preBPix": "trg_double_tau35_mediumiso_hps",
-                            "2023postBPix": "trg_double_tau35_mediumiso_hps",
-                            "2024": "trg_double_tau30_mediumiso_pnet",
-                            "2025": "trg_double_tau30_mediumiso_pnet",
-                            "2026": "trg_double_tau30_mediumiso_pnet",
-                        }
-                    ),
-                    "doubletau_trigger_leg1_sf_name": EraModifier(
-                        {
-                            "2016preVFP": '""',
-                            "2016postVFP": '""', 
-                            "2017": '""',
-                            "2018": '""',
-                            "2022preEE":"ditau",
-                            "2022postEE":"ditau",
-                            "2023preBPix":"ditau",
-                            "2023postBPix":"ditau",
-                            "2024":"ditau",
-                            "2025":"ditau",
-                            "2026":"ditau",
-                        }
-                    ),
+                    "doubletau_trigger_leg1_flagname": DOUBLETAU_TRIGGER_LEG1_FLAGNAME,
+                    "doubletau_trigger_flag": DOUBLETAU_TRIGGER_FLAG,
+                    "doubletau_trigger_leg1_sf_name": DOUBLETAU_TRIGGER_SF_NAME,
                     "doubletau_trigger_leg1_variation": "nom",
                 },
             ],
             "doubletau_trigger_leg2_sf": [
                 {
-                    "doubletau_trigger_leg2_flagname": EraModifier(
-                        {
-                            "2016preVFP": '""',
-                            "2016postVFP": '""', 
-                            "2017": '""',
-                            "2018": '""',
-                            "2022preEE": "trg_wgt_doubletau35_leg2",
-                            "2022postEE": "trg_wgt_doubletau35_leg2",
-                            "2023preBPix": "trg_wgt_doubletau35_leg2",
-                            "2023postBPix": "trg_wgt_doubletau35_leg2",
-                            "2024": "trg_wgt_doubletau30_leg2",
-                            "2025": "trg_wgt_doubletau30_leg2",
-                            "2026": "trg_wgt_doubletau30_leg2",
-                        }
-                    ),
-                    "doubletau_trigger_flag": EraModifier(
-                        {
-                            "2016preVFP": '""',
-                            "2016postVFP": '""', 
-                            "2017": '""',
-                            "2018": '""',
-                            "2022preEE": "trg_double_tau35_mediumiso_hps",
-                            "2022postEE": "trg_double_tau35_mediumiso_hps",
-                            "2023preBPix": "trg_double_tau35_mediumiso_hps",
-                            "2023postBPix": "trg_double_tau35_mediumiso_hps",
-                            "2024": "trg_double_tau30_mediumiso_pnet",
-                            "2025": "trg_double_tau30_mediumiso_pnet",
-                            "2026": "trg_double_tau30_mediumiso_pnet",
-                        }
-                    ),
-                    "doubletau_trigger_leg2_sf_name": EraModifier(
-                        {
-                            "2016preVFP": '""',
-                            "2016postVFP": '""', 
-                            "2017": '""',
-                            "2018": '""',
-                            "2022preEE":"ditau",
-                            "2022postEE":"ditau",
-                            "2023preBPix":"ditau",
-                            "2023postBPix":"ditau",
-                            "2024":"ditau",
-                            "2025":"ditau",
-                            "2026":"ditau",
-                        }
-                    ),
+                    "doubletau_trigger_leg2_flagname": DOUBLETAU_TRIGGER_LEG2_FLAGNAME,
+                    "doubletau_trigger_flag": DOUBLETAU_TRIGGER_FLAG,
+                    "doubletau_trigger_leg2_sf_name": DOUBLETAU_TRIGGER_SF_NAME,
                     "doubletau_trigger_leg2_variation": "nom",
                 },
             ],
         },
     )
-
-    #######################
-    # trigger shifts #
-    #######################
-
-    for _variation in ["up", "down"]:
-        configuration.add_shift(
-            SystematicShift(
-                name=f"CMS_eff_e_trigger{_variation.upper()}",
-                shift_config={
-                    ("et"): {
-                        "singleelectron_trigger_sf": [
-                            {
-                                "singleelectron_trigger_flagname": "trg_wgt_single_ele30",
-                                "singleelectron_trigger_flag": "trg_single_ele30",
-                                "singleelectron_trigger_sf_name": "Electron-HLT-SF",
-                                "singleelectron_trigger_path_id_name": "HLT_SF_Ele30_MVAiso90ID",
-                                "singleelectron_trigger_variation": f"sf{_variation}",
-                            },
-                        ],
-                    }
-                },
-                producers={("et"): scalefactors.SingleEleTriggerSF},
-            ),
-            exclude_samples=["data", "embedding", "embedding_mc"],
-        )
-        configuration.add_shift(
-            SystematicShift(
-                name=f"CMS_trig_etau_cross{_variation.upper()}",
-                shift_config={
-                    ("et"): {
-                        "eletau_cross_trigger_leg1_sf": [
-                            {
-                                "eletau_cross_trigger_leg1_flagname": "trg_wgt_ele24tau30_leg1",
-                                "eletau_cross_trigger_flag": EraModifier(
-                                    {
-                                        "2016preVFP": "trg_cross_ele24tau30_hps",
-                                        "2016postVFP": "trg_cross_ele24tau30_hps",
-                                        "2017": "trg_cross_ele24tau30_hps",
-                                        "2018": "trg_cross_ele24tau30_hps",
-                                        "2022preEE": "trg_cross_ele24tau30_hps",
-                                        "2022postEE": "trg_cross_ele24tau30_hps",
-                                        "2023preBPix": "trg_cross_ele24tau30_hps",
-                                        "2023postBPix": "trg_cross_ele24tau30_hps",
-                                        "2024": "trg_cross_ele24tau30_hps",
-                                        "2025": "trg_cross_ele24tau30_pnet",
-                                        "2026": "trg_cross_ele24tau30_pnet",
-                                    }
-                                ),
-                                "eletau_cross_trigger_leg1_sf_name": "Electron-HLT-SF",
-                                "eletau_cross_trigger_leg1_path_id_name": "HLT_SF_Ele24_TightID",
-                                "eletau_cross_trigger_leg1_variation": f"sf{_variation}",
-                            },
-                        ],
-                        "eletau_cross_trigger_leg2_sf": [
-                            {
-                                "eletau_cross_trigger_leg2_flagname": "trg_wgt_ele24tau30_leg2",
-                                "eletau_cross_trigger_flag": EraModifier(
-                                    {
-                                        "2016preVFP": "trg_cross_ele24tau30_hps",
-                                        "2016postVFP": "trg_cross_ele24tau30_hps",
-                                        "2017": "trg_cross_ele24tau30_hps",
-                                        "2018": "trg_cross_ele24tau30_hps",
-                                        "2022preEE": "trg_cross_ele24tau30_hps",
-                                        "2022postEE": "trg_cross_ele24tau30_hps",
-                                        "2023preBPix": "trg_cross_ele24tau30_hps",
-                                        "2023postBPix": "trg_cross_ele24tau30_hps",
-                                        "2024": "trg_cross_ele24tau30_hps",
-                                        "2025": "trg_cross_ele24tau30_pnet",
-                                        "2026": "trg_cross_ele24tau30_pnet",
-                                    }
-                                ),
-                                "eletau_cross_trigger_leg2_sf_name": "etau",
-                                "eletau_cross_trigger_leg2_variation": _variation,
-                            },
-                        ]
-                    },
-                },
-                producers={
-                    ("et"): [
-                        scalefactors.EleTauTriggerSF,
-                    ],
-                },
-            ),
-            exclude_samples=["data", "embedding", "embedding_mc"],
-        )
-        configuration.add_shift(
-                SystematicShift(
-                    name=f"CMS_eff_m_trigger{_variation.upper()}",
-                    shift_config={
-                        ("mt"): {
-                            "singlemuon_trigger_sf": [
-                                {
-                                    "singlemuon_trigger_flagname": "trg_wgt_single_mu24",
-                                    "singlemuon_trigger_flag": "trg_single_mu24",
-                                    "singlemuon_trigger_sf_name": "NUM_IsoMu24_DEN_CutBasedIdTight_and_PFIsoTight",
-                                    "singlemuon_trigger_variation": f"syst{_variation}",
-                                },
-                            ],
-                        }
-                    },
-                    producers={("mt"): scalefactors.SingleMuTriggerSF},
-                ),
-                exclude_samples=["data", "embedding", "embedding_mc"],
-            )
-        configuration.add_shift(
-            SystematicShift(
-                name=f"CMS_trig_mutau_cross{_variation.upper()}",
-                shift_config={
-                    ("mt"): {
-                        "mutau_trigger_leg1_sf": [
-                            {
-                                "mutau_cross_trigger_flag": EraModifier(
-                                    {
-                                        "2016preVFP": "trg_cross_mu20tau27_hps",
-                                        "2016postVFP": "trg_cross_mu20tau27_hps",
-                                        "2017": "trg_cross_mu20tau27_hps",
-                                        "2018": "trg_cross_mu20tau27_hps",
-                                        "2022preEE": "trg_cross_mu20tau27_hps",
-                                        "2022postEE": "trg_cross_mu20tau27_hps",
-                                        "2023preBPix": "trg_cross_mu20tau27_hps",
-                                        "2023postBPix": "trg_cross_mu20tau27_hps",
-                                        "2024": "trg_cross_mu20tau27_hps",
-                                        "2025": "trg_cross_mu20tau27_pnet",
-                                        "2026": "trg_cross_mu20tau27_pnet",
-                                    }
-                                ),
-                                "mutau_cross_trigger_leg1_flagname": "trg_wgt_mu20tau27_leg1",
-                                "mutau_cross_trigger_leg1_sf_name": "NUM_IsoMu20_DEN_CutBasedIdTight_and_PFIsoTight",
-                                "mutau_cross_trigger_leg1_variation": f"syst{_variation}",
-                            },
-                        ],
-                        "mutau_trigger_leg2_sf": [
-                            {
-                                "mutau_cross_trigger_flag": EraModifier(
-                                    {
-                                        "2016preVFP": "trg_cross_mu20tau27_hps",
-                                        "2016postVFP": "trg_cross_mu20tau27_hps",
-                                        "2017": "trg_cross_mu20tau27_hps",
-                                        "2018": "trg_cross_mu20tau27_hps",
-                                        "2022preEE": "trg_cross_mu20tau27_hps",
-                                        "2022postEE": "trg_cross_mu20tau27_hps",
-                                        "2023preBPix": "trg_cross_mu20tau27_hps",
-                                        "2023postBPix": "trg_cross_mu20tau27_hps",
-                                        "2024": "trg_cross_mu20tau27_hps",
-                                        "2025": "trg_cross_mu20tau27_pnet",
-                                        "2026": "trg_cross_mu20tau27_pnet",
-                                    }
-                                ),
-                                "mutau_cross_trigger_leg2_flagname": "trg_wgt_mu20tau27_leg2",
-                                "mutau_cross_trigger_leg2_sf_name": "mutau",
-                                "mutau_cross_trigger_leg2_variation": _variation,
-                            },
-                        ],
-                    },
-                },
-                producers={
-                    ("mt"): [
-                        scalefactors.MuTauTriggerSF,
-                    ],
-                },
-            ),
-            exclude_samples=["data", "embedding", "embedding_mc"],
-        )
-        configuration.add_shift(
-            SystematicShift(
-                name=f"CMS_trig_t_ditau_cross{_variation.upper()}",
-                shift_config={
-                    ("tt"): {
-                        "doubletau_trigger_leg1_sf": [
-                            {
-                                "doubletau_trigger_leg1_flagname": EraModifier(
-                                    {
-                                        "2016preVFP": '""',
-                                        "2016postVFP": '""', 
-                                        "2017": '""',
-                                        "2018": '""',
-                                        "2022preEE": "trg_wgt_doubletau35_leg1",
-                                        "2022postEE": "trg_wgt_doubletau35_leg1",
-                                        "2023preBPix": "trg_wgt_doubletau35_leg1",
-                                        "2023postBPix": "trg_wgt_doubletau35_leg1",
-                                        "2024": "trg_wgt_doubletau30_leg1",
-                                        "2025": "trg_wgt_doubletau30_leg1",
-                                        "2026": "trg_wgt_doubletau30_leg1",
-                                    }
-                                ),
-                                "doubletau_trigger_flag":  EraModifier(
-                                    {
-                                        "2016preVFP": '""',
-                                        "2016postVFP": '""', 
-                                        "2017": '""',
-                                        "2018": '""',
-                                        "2022preEE": "trg_double_tau35_mediumiso_hps",
-                                        "2022postEE": "trg_double_tau35_mediumiso_hps",
-                                        "2023preBPix": "trg_double_tau35_mediumiso_hps",
-                                        "2023postBPix": "trg_double_tau35_mediumiso_hps",
-                                        "2024": "trg_double_tau30_mediumiso_pnet",
-                                        "2025": "trg_double_tau30_mediumiso_pnet",
-                                        "2026": "trg_double_tau30_mediumiso_pnet",
-                                    }
-                                ),
-                                "doubletau_trigger_leg1_sf_name": EraModifier(
-                                    {
-                                        "2016preVFP": '""',
-                                        "2016postVFP": '""', 
-                                        "2017": '""',
-                                        "2018": '""',
-                                        "2022preEE":"ditau",
-                                        "2022postEE":"ditau",
-                                        "2023preBPix":"ditau",
-                                        "2023postBPix":"ditau",
-                                        "2024":"ditau",
-                                        "2025":"ditau",
-                                        "2026":"ditau",
-                                    }
-                                ),
-                                "doubletau_trigger_leg1_variation": _variation,
-                            },
-                        ],
-                        "doubletau_trigger_leg2_sf": [
-                            {
-                                "doubletau_trigger_leg2_flagname": EraModifier(
-                                    {
-                                        "2016preVFP": '""',
-                                        "2016postVFP": '""', 
-                                        "2017": '""',
-                                        "2018": '""',
-                                        "2022preEE": "trg_wgt_doubletau35_leg2",
-                                        "2022postEE": "trg_wgt_doubletau35_leg2",
-                                        "2023preBPix": "trg_wgt_doubletau35_leg2",
-                                        "2023postBPix": "trg_wgt_doubletau35_leg2",
-                                        "2024": "trg_wgt_doubletau30_leg2",
-                                        "2025": "trg_wgt_doubletau30_leg2",
-                                        "2026": "trg_wgt_doubletau30_leg2",
-                                    }
-                                ),
-                                "doubletau_trigger_flag":  EraModifier(
-                                    {
-                                        "2016preVFP": '""',
-                                        "2016postVFP": '""', 
-                                        "2017": '""',
-                                        "2018": '""',
-                                        "2022preEE": "trg_double_tau35_mediumiso_hps",
-                                        "2022postEE": "trg_double_tau35_mediumiso_hps",
-                                        "2023preBPix": "trg_double_tau35_mediumiso_hps",
-                                        "2023postBPix": "trg_double_tau35_mediumiso_hps",
-                                        "2024": "trg_double_tau30_mediumiso_pnet",
-                                        "2025": "trg_double_tau30_mediumiso_pnet",
-                                        "2026": "trg_double_tau30_mediumiso_pnet",
-                                    }
-                                ),
-                                "doubletau_trigger_leg2_sf_name": EraModifier(
-                                    {
-                                        "2016preVFP": '""',
-                                        "2016postVFP": '""', 
-                                        "2017": '""',
-                                        "2018": '""',
-                                        "2022preEE":"ditau",
-                                        "2022postEE":"ditau",
-                                        "2023preBPix":"ditau",
-                                        "2023postBPix":"ditau",
-                                        "2024":"ditau",
-                                        "2025":"ditau",
-                                        "2026":"ditau",
-                                    }
-                                ),
-                                "doubletau_trigger_leg2_variation": _variation,
-                            },
-                        ],
-                    },
-                },
-                producers={
-                    ("tt"): [
-                        scalefactors.DoubleTauTriggerSF,
-                    ],
-                },
-            ),
-            exclude_samples=["data", "embedding", "embedding_mc"],
-        )
 
     #####################
     ####### RUN 2 #######
@@ -2005,104 +1135,44 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
         {
             "singlemuon_trigger_sf_mc": EraModifier(
                 {
-                    "2026": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2025": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2024": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2023postBPix": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2023preBPix": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2022postEE": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2022preEE": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2018": [
-                        {
-                            "flagname": "trg_wgt_single_mu24",
-                            "mc_trigger_sf": "Trg_IsoMu24_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                        {
-                            "flagname": "trg_wgt_single_mu27",
-                            "mc_trigger_sf": "Trg_IsoMu27_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                        {
-                            "flagname": "trg_wgt_single_mu24ormu27",
-                            "mc_trigger_sf": "Trg_IsoMu27_or_IsoMu24_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2017": [
-                        {
-                            "flagname": "trg_wgt_single_mu24",
-                            "mc_trigger_sf": "Trg_IsoMu24_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                        {
-                            "flagname": "trg_wgt_single_mu27",
-                            "mc_trigger_sf": "Trg_IsoMu27_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                        {
-                            "flagname": "trg_wgt_single_mu24ormu27",
-                            "mc_trigger_sf": "Trg_IsoMu27_or_IsoMu24_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2016postVFP": [
-                        {
-                            "flagname": "trg_wgt_single_mu22",
-                            "mc_trigger_sf": "Trg_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2016preVFP": [
-                        {
-                            "flagname": "trg_wgt_single_mu22",
-                            "mc_trigger_sf": "Trg_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                }
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_wgt_single_mu24",
+                                "mc_trigger_sf": "Trg_IsoMu24_pt_eta_bins",
+                                "mc_trg_extrapolation": 1.0,  # for nominal case
+                            },
+                            {
+                                "flagname": "trg_wgt_single_mu27",
+                                "mc_trigger_sf": "Trg_IsoMu27_pt_eta_bins",
+                                "mc_trg_extrapolation": 1.0,  # for nominal case
+                            },
+                            {
+                                "flagname": "trg_wgt_single_mu24ormu27",
+                                "mc_trigger_sf": "Trg_IsoMu27_or_IsoMu24_pt_eta_bins",
+                                "mc_trg_extrapolation": 1.0,  # for nominal case
+                            },
+                        ]
+                        for era in ["2017", "2018"]
+                    },
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_wgt_single_mu22",
+                                "mc_trigger_sf": "Trg_pt_eta_bins",
+                                "mc_trg_extrapolation": 1.0,  # for nominal case
+                            },
+                        ]
+                        for era in ["2016preVFP", "2016postVFP"]
+                    },
+                },
+                default=[  ## TODO: not implemented, here as a placeholder (2022preEE, 2022postEE, 2023preBPix, 2023postBPix, 2024, 2025, 2026)
+                    {
+                        "flagname": '""',
+                        "mc_trigger_sf": '""',
+                        "mc_trg_extrapolation": 1.0,  # for nominal case
+                    },
+                ],
             )
         },
     )
@@ -2112,590 +1182,51 @@ def add_diTauTriggerSetup(configuration: Configuration) -> Configuration:
         {
             "singlelectron_trigger_sf_mc": EraModifier(
                 {
-                    "2026": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2025": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2024": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2023postBPix": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2023preBPix": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2022postEE": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2022preEE": [ ## TODO: not implemented, here as a placeholder
-                        {
-                            "flagname": '""',
-                            "mc_trigger_sf": '""',
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2018": [
-                        {
-                            "flagname": "trg_wgt_single_ele32",
-                            "mc_trigger_sf": "Trg32_Iso_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                        {
-                            "flagname": "trg_wgt_single_ele35",
-                            "mc_trigger_sf": "Trg35_Iso_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                        {
-                            "flagname": "trg_wgt_single_ele32orele35",
-                            "mc_trigger_sf": "Trg32_or_Trg35_Iso_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                        {
-                            "flagname": "trg_wgt_single_ele27orele32orele35",
-                            "mc_trigger_sf": "Trg_Iso_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2017": [
-                        {
-                            "flagname": "trg_wgt_single_ele32",
-                            "mc_trigger_sf": "Trg32_Iso_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                        {
-                            "flagname": "trg_wgt_single_ele35",
-                            "mc_trigger_sf": "Trg35_Iso_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                        {
-                            "flagname": "trg_wgt_single_ele32orele35",
-                            "mc_trigger_sf": "Trg32_or_Trg35_Iso_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                        {
-                            "flagname": "trg_wgt_single_ele27orele32orele35",
-                            "mc_trigger_sf": "Trg_Iso_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        },
-                    ],
-                    "2016postVFP": [
-                        {
-                            "flagname": "trg_wgt_single_ele25",
-                            "mc_trigger_sf": "Trg25_Iso_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        }
-                    ],
-                    "2016preVFP": [
-                        {
-                            "flagname": "trg_wgt_single_ele25",
-                            "mc_trigger_sf": "Trg25_Iso_pt_eta_bins",
-                            "mc_trg_extrapolation": 1.0,  # for nominal case
-                        }
-                    ],
-                }
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_wgt_single_ele32",
+                                "mc_trigger_sf": "Trg32_Iso_pt_eta_bins",
+                                "mc_trg_extrapolation": 1.0,  # for nominal case
+                            },
+                            {
+                                "flagname": "trg_wgt_single_ele35",
+                                "mc_trigger_sf": "Trg35_Iso_pt_eta_bins",
+                                "mc_trg_extrapolation": 1.0,  # for nominal case
+                            },
+                            {
+                                "flagname": "trg_wgt_single_ele32orele35",
+                                "mc_trigger_sf": "Trg32_or_Trg35_Iso_pt_eta_bins",
+                                "mc_trg_extrapolation": 1.0,  # for nominal case
+                            },
+                            {
+                                "flagname": "trg_wgt_single_ele27orele32orele35",
+                                "mc_trigger_sf": "Trg_Iso_pt_eta_bins",
+                                "mc_trg_extrapolation": 1.0,  # for nominal case
+                            },
+                        ]
+                        for era in ["2017", "2018"]
+                    },
+                    **{
+                        era: [
+                            {
+                                "flagname": "trg_wgt_single_ele25",
+                                "mc_trigger_sf": "Trg25_Iso_pt_eta_bins",
+                                "mc_trg_extrapolation": 1.0,  # for nominal case
+                            }
+                        ]
+                        for era in ["2016preVFP", "2016postVFP"]
+                    },
+                },
+                default=[  ## TODO: not implemented, here as a placeholder (2022preEE, 2022postEE, 2023preBPix, 2023postBPix, 2024, 2025, 2026)
+                    {
+                        "flagname": '""',
+                        "mc_trigger_sf": '""',
+                        "mc_trg_extrapolation": 1.0,  # for nominal case
+                    },
+                ],
             )
         },
-    )
-
-    ##################
-    # trigger shifts #
-    ##################
-
-    configuration.add_shift(
-        SystematicShift(
-            name="singleElectronTriggerSFUp",
-            shift_config={
-                ("et"): {
-                    "singlelectron_trigger_sf_mc": EraModifier(
-                        {
-                            "2026": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2025": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2024": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2023postBPix": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2023preBPix": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2022postEE": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2022preEE": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2018": [
-                                {
-                                    "flagname": "trg_wgt_single_ele32orele35",
-                                    "mc_trigger_sf": "Trg32_or_Trg35_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_ele32",
-                                    "mc_trigger_sf": "Trg32_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_ele35",
-                                    "mc_trigger_sf": "Trg35_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_ele27orele32orele35",
-                                    "mc_trigger_sf": "Trg_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                            ],
-                            "2017": [
-                                {
-                                    "flagname": "trg_wgt_single_ele32orele35",
-                                    "mc_trigger_sf": "Trg32_or_Trg35_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_ele32",
-                                    "mc_trigger_sf": "Trg32_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_ele35",
-                                    "mc_trigger_sf": "Trg35_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_ele27orele32orele35",
-                                    "mc_trigger_sf": "Trg_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                            ],
-                            "2016postVFP": [
-                                {
-                                    "flagname": "trg_wgt_single_ele25",
-                                    "mc_trigger_sf": "Trg25_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                }
-                            ],
-                            "2016preVFP": [
-                                {
-                                    "flagname": "trg_wgt_single_ele25",
-                                    "mc_trigger_sf": "Trg25_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                }
-                            ],
-                        }
-                    )
-                }
-            },
-            producers={("et"): scalefactors.ETGenerateSingleElectronTriggerSF_MC},
-        ),
-        exclude_samples=["data", "embedding", "embedding_mc"],
-    )
-    configuration.add_shift(
-        SystematicShift(
-            name="singleElectronTriggerSFDown",
-            shift_config={
-                ("et"): {
-                    "singlelectron_trigger_sf_mc": EraModifier(
-                        {
-                            "2026": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2025": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2024": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2023postBPix": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2023preBPix": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2022postEE": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2022preEE": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2018": [
-                                {
-                                    "flagname": "trg_wgt_single_ele32orele35",
-                                    "mc_trigger_sf": "Trg32_or_Trg35_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_ele32",
-                                    "mc_trigger_sf": "Trg32_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_ele35",
-                                    "mc_trigger_sf": "Trg35_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_ele27orele32orele35",
-                                    "mc_trigger_sf": "Trg_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                            ],
-                            "2017": [
-                                {
-                                    "flagname": "trg_wgt_single_ele32orele35",
-                                    "mc_trigger_sf": "Trg32_or_Trg35_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_ele32",
-                                    "mc_trigger_sf": "Trg32_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_ele35",
-                                    "mc_trigger_sf": "Trg35_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_ele27orele32orele35",
-                                    "mc_trigger_sf": "Trg_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                            ],
-                            "2016postVFP": [
-                                {
-                                    "flagname": "trg_wgt_single_ele25",
-                                    "mc_trigger_sf": "Trg25_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                }
-                            ],
-                            "2016preVFP": [
-                                {
-                                    "flagname": "trg_wgt_single_ele25",
-                                    "mc_trigger_sf": "Trg25_Iso_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                }
-                            ],
-                        }
-                    )
-                }
-            },
-            producers={("et"): scalefactors.ETGenerateSingleElectronTriggerSF_MC},
-        ),
-        exclude_samples=["data", "embedding", "embedding_mc"],
-    )
-
-    configuration.add_shift(
-        SystematicShift(
-            name="singleMuonTriggerSFUp",
-            shift_config={
-                ("mt"): {
-                    "singlemuon_trigger_sf_mc": EraModifier(
-                        {
-                            "2026": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2025": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2024": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2023postBPix": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2023preBPix": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2022postEE": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2022preEE": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2018": [
-                                {
-                                    "flagname": "trg_wgt_single_mu24",
-                                    "mc_trigger_sf": "Trg_IsoMu24_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_mu27",
-                                    "mc_trigger_sf": "Trg_IsoMu27_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_mu24ormu27",
-                                    "mc_trigger_sf": "Trg_IsoMu27_or_IsoMu24_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                            ],
-                            "2017": [
-                                {
-                                    "flagname": "trg_wgt_single_mu24",
-                                    "mc_trigger_sf": "Trg_IsoMu24_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_mu27",
-                                    "mc_trigger_sf": "Trg_IsoMu27_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_mu24ormu27",
-                                    "mc_trigger_sf": "Trg_IsoMu27_or_IsoMu24_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                            ],
-                            "2016postVFP": [
-                                {
-                                    "flagname": "trg_wgt_single_mu22",
-                                    "mc_trigger_sf": "Trg_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                            ],
-                            "2016preVFP": [
-                                {
-                                    "flagname": "trg_wgt_single_mu22",
-                                    "mc_trigger_sf": "Trg_pt_eta_bins",
-                                    "mc_trg_extrapolation": 1.02,
-                                },
-                            ],
-                        }
-                    )
-                }
-            },
-            producers={("mt"): scalefactors.MTGenerateSingleMuonTriggerSF_MC},
-        ),
-        exclude_samples=["data", "embedding", "embedding_mc"],
-    )
-    configuration.add_shift(
-        SystematicShift(
-            name="singleMuonTriggerSFDown",
-            shift_config={
-                ("mt"): {
-                    "singlemuon_trigger_sf_mc": EraModifier(
-                        {
-                            "2026": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2025": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2024": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2023postBPix": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2023preBPix": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2022postEE": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2022preEE": [ ## TODO: not implemented, here as a placeholder
-                                {
-                                    "flagname": '""',
-                                    "mc_trigger_sf": '""',
-                                    "mc_trg_extrapolation": 1.0,  # for nominal case
-                                },
-                            ],
-                            "2018": [
-                                {
-                                    "flagname": "trg_wgt_single_mu24",
-                                    "mc_trigger_sf": "Trg_IsoMu24_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_mu27",
-                                    "mc_trigger_sf": "Trg_IsoMu27_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_mu24ormu27",
-                                    "mc_trigger_sf": "Trg_IsoMu27_or_IsoMu24_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                            ],
-                            "2017": [
-                                {
-                                    "flagname": "trg_wgt_single_mu24",
-                                    "mc_trigger_sf": "Trg_IsoMu24_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_mu27",
-                                    "mc_trigger_sf": "Trg_IsoMu27_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                                {
-                                    "flagname": "trg_wgt_single_mu24ormu27",
-                                    "mc_trigger_sf": "Trg_IsoMu27_or_IsoMu24_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                            ],
-                            "2016postVFP": [
-                                {
-                                    "flagname": "trg_wgt_single_mu22",
-                                    "mc_trigger_sf": "Trg_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                            ],
-                            "2016preVFP": [
-                                {
-                                    "flagname": "trg_wgt_single_mu22",
-                                    "mc_trigger_sf": "Trg_pt_eta_bins",
-                                    "mc_trg_extrapolation": 0.98,
-                                },
-                            ],
-                        }
-                    )
-                }
-            },
-            producers={("mt"): scalefactors.MTGenerateSingleMuonTriggerSF_MC},
-        ),
-        exclude_samples=["data", "embedding", "embedding_mc"],
     )
 
     return configuration
